% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{all_dataset}
\alias{all_dataset}
\title{列举所有数据集}
\usage{
all_dataset(topic = "CACHE")
}
\description{
列举所有数据集
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{last_affected_parts}()},
\code{\link{read_affected_parts}()},
\code{\link{read_dataset}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
