% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_task_queue_create}
\alias{build_task_queue_create}
\title{创建BUILD任务批处理队列}
\usage{
build_task_queue_create(
  taskQueue = "__TASK_QUEUE__",
  taskTopic = "TASK_DEFINE",
  cacheTopic = "CACHE"
)
}
\description{
创建BUILD任务批处理队列
}
\seealso{
Other build function: 
\code{\link{build_task_queue_run}()}
}
\concept{build function}
