% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_load}
\alias{config_load}
\title{加载配置文件到内存}
\usage{
config_load(path = "./", yml = "config.yml")
}
\arguments{
\item{path}{默认为配置文件所在的目录}

\item{yml}{默认为config.yml}
}
\description{
自动创建\code{ROOT_PATH}目录。

根据配置文件配置目录到运行环境。
}
\details{
使用\code{ROOT_PATH}时有一个关键约定：
配置项必须以\code{./}开头，才能使用\code{ROOT_PATH}扩展其路径；
否则，将被视为独立配置名。
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_write}()},
\code{\link{config_yaml}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()}
}
\concept{config functions}
