% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_write}
\alias{config_write}
\title{创建或补写配置项}
\usage{
config_write(path, yml = "config.yml", option = list())
}
\arguments{
\item{path}{YAML配置文件目录}

\item{yml}{默认为config.yml}

\item{option}{配置项}
}
\description{
多次运行时会增量补充
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_yaml}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()}
}
\concept{config functions}
