% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_yaml}
\alias{config_yaml}
\title{读取配置项文件}
\usage{
config_yaml(path = "./", yml = "config.yml")
}
\arguments{
\item{path}{YAML配置文件目录}

\item{yml}{默认为config.yml}
}
\description{
读取配置项文件
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()}
}
\concept{config functions}
