% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir}
\alias{create_dir}
\title{创建文件夹}
\usage{
create_dir(path)
}
\description{
创建文件夹
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%error\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{gen_batchNum}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
