% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_arrange}
\alias{dp_arrange}
\title{行排序}
\usage{
dp_arrange(
  sn_columns = list(),
  b_desc = FALSE,
  b_by_group = FALSE,
  s_dataName = "@result"
)
}
\description{
行排序
}
\seealso{
Other data-plyr function: 
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_rename}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
