% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_collect}
\alias{dp_collect}
\title{立即执行数据收集（结束惰性计算）}
\usage{
dp_collect(dataName = "@result")
}
\description{
立即执行数据收集（结束惰性计算）
}
\seealso{
Other data-plyr function: 
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
