% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_head}
\alias{dp_head}
\title{头部数据}
\usage{
dp_head(n = 10, dataName = "@result")
}
\description{
头部数据
}
\seealso{
Other data-plyr function: 
\code{\link{dp_arrange}()},
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_rename}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
