% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_n_min}
\alias{dp_n_min}
\title{按最小值取N条记录}
\usage{
dp_n_min(orderColumn, n = 10, with_ties = FALSE, dataName = "@result")
}
\description{
按最小值取N条记录
}
\seealso{
Other data-plyr function: 
\code{\link{dp_arrange}()},
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_read}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
