% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_read}
\alias{dp_read}
\title{读取数据集}
\usage{
dp_read(dsName, dataName = "@result")
}
\description{
读取数据集
}
\seealso{
Other data-plyr function: 
\code{\link{dp_arrange}()},
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
