% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_rename}
\alias{dp_rename}
\title{列改名}
\usage{
dp_rename(newName, oldName, dataName = "@result")
}
\description{
列改名
}
\seealso{
Other data-plyr function: 
\code{\link{dp_arrange}()},
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
