% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_select}
\alias{dp_select}
\title{按最小值取N条记录}
\usage{
dp_select(
  columns = list(),
  everything = FALSE,
  regex = NULL,
  dataName = "@result"
)
}
\description{
按最小值取N条记录
}
\seealso{
Other data-plyr function: 
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
