% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_tail}
\alias{dp_tail}
\title{尾部数据}
\usage{
dp_tail(n = 10, dataName = "@result")
}
\description{
尾部数据
}
\seealso{
Other data-plyr function: 
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_select}()}
}
\concept{data-plyr function}
