% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_add_count}
\alias{ds_add_count}
\title{计数统计，并将结果追加到原数据集}
\usage{
ds_add_count(d, columns = c(), sort = FALSE, name = "n")
}
\description{
计数统计，并将结果追加到原数据集
}
\seealso{
Other gali-dataset function: 
\code{\link{ds_arrange}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()}
}
\concept{gali-dataset function}
