% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_all}
\alias{ds_all}
\title{列举所有数据集}
\usage{
ds_all(topic = "CACHE")
}
\arguments{
\item{topic}{主题域}
}
\description{
列举所有数据集
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_last_affected}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()},
\code{\link{ds_yaml}()}
}
\concept{dataset function}
