% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_arrange}
\alias{ds_arrange}
\title{行排序}
\usage{
ds_arrange(d, columns = list(), desc = FALSE, by_group = FALSE)
}
\description{
行排序
}
\seealso{
Other gali-dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()}
}
\concept{gali-dataset function}
