% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_as_from}
\alias{ds_as_from}
\title{设置数据来源}
\usage{
ds_as_from(ds, sourceFrom)
}
\description{
设置数据来源
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_datetime}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_drop}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
