% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_as_unique}
\alias{ds_as_unique}
\title{去除重复行}
\usage{
ds_as_unique(ds, keyColumns)
}
\arguments{
\item{ds}{要确认的数据集}

\item{columns}{要确认的列名或其向量、列表}
}
\description{
按键值列去除重复行，保留最先发现的行
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
