% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_collect}
\alias{ds_collect}
\title{立即执行数据收集（结束惰性计算）}
\usage{
ds_collect(d)
}
\description{
立即执行数据收集（结束惰性计算）
}
\seealso{
Other gali-dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_arrange}()},
\code{\link{ds_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()}
}
\concept{gali-dataset function}
