% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{ds_compare_schema}
\alias{ds_compare_schema}
\title{确认数据集同构}
\usage{
ds_compare_schema(schema1, schema2)
}
\arguments{
\item{schema1}{用于比较的数据集结构，可使用ds_schema获得}

\item{schema2}{用于参考的数据集结构，可使用ds_schema获得}
}
\description{
比较两个数据集的列数、列名、列字段类型是否一致。
}
\details{
如果列结构完全相等，则返回列表中的equal为TRUE，否则为FALSE；
如果第一个数据集中的列在第二个数据集中全部存在，且字段类型一致，则返回列表中的
contained为TRUE，否则为FALSE。

在需要时，可以将ds_schema获得的数据集结构持久化保存（例如，使用saveRDS），
然后用于比较新导入、新生成的数据集是否合规。
也可以直接使用 dataset_comfirm 来完成这个工作。
}
\seealso{
Other dataset function: 
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{read_dataset}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
