% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_delete}
\alias{ds_delete}
\title{批量追加删除的数据}
\usage{
ds_delete(d, dsName, topic = "CACHE")
}
\arguments{
\item{d}{要删除的数据}

\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}
}
\description{
以追加标记文件的形式删除数据（不做物理删除）
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
