% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_diff_dataset}
\alias{ds_diff_dataset}
\title{详细比较两个数据集结构}
\usage{
ds_diff_dataset(ds1, ds2)
}
\arguments{
\item{ds1}{用于比较的数据集结构，可使用\code{\link{ds_schema}}获得}

\item{ds2}{用于参考的数据集结构，可使用\code{\link{ds_schema}}获得}
}
\description{
详细比较两个数据集结构
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_init}()},
\code{\link{ds_last_affected}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
