% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-expand.R
\name{ds_expand_init}
\alias{ds_expand_init}
\title{初始化扩展数据集}
\usage{
ds_expand_init(..., bindDataset)
}
\arguments{
\item{...}{参考函数 \link{ds_init}}

\item{bindDataset}{绑定的数据集}
}
\description{
对数据集实现扩展列或数据标注。
}
\concept{dataset-expand function}
