% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_files}
\alias{ds_files}
\title{读取数据集的文件信息}
\usage{
ds_files(dsName, topic = "CACHE")
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}
}
\description{
读取数据集的文件信息
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_init}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
