% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_filter}
\alias{ds_filter}
\title{定义数据过滤任务}
\usage{
ds_filter(d, column, op, value = list(NULL))
}
\arguments{
\item{column}{列名}

\item{op}{阈值范围判断符号}

\item{value}{阈值}

\item{s_dataName}{内存中的数据框名称，默认为 @result}

\item{taskTopic}{任务定义的主题文件夹}
}
\description{
这个任务的目标是通过配置项实现函数\code{\link[dplyr:filter]{dplyr::filter()}}的大部分功能。

允许为数据集增加多个阈值查询条件，缩小筛查范围。

taskId、column、op、value等参数构造唯一的gali_ds_filter表达式，
这将允许从UI生成或还原该操作。

只要包括以下逻辑判断：
\itemize{
\item >, <, >=, <=, ==, !=
\item \%in\%, \%nin\%
\item \%regex\%, \%not-regex\%
\item #time# >, #time# <
\item #date# >, #date# <
}
}
\seealso{
Other dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_arrange}()},
\code{\link{ds_as_datetime}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_delete}()},
\code{\link{ds_demo}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_distinct}()},
\code{\link{ds_drop}()},
\code{\link{ds_dup_count}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_head}()},
\code{\link{ds_init}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_rename}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
