% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-import.R
\name{ds_import_init}
\alias{ds_import_init}
\title{初始化导入数据集}
\usage{
ds_import_init(
  ...,
  importName,
  importTopic = "IMPORT",
  taskName,
  taskTopic = "TASK"
)
}
\arguments{
\item{...}{参考函数 \link{ds_init}}

\item{importName}{导入素材所在的文件夹名称}

\item{importTopic}{导入素材主题文件夹位置}

\item{taskName}{任务脚本名称}

\item{taskTopic}{任务脚本主题文件夹位置}
}
\description{
导入配置涉及到输入素材、输出结果，以及处理队列和处理脚本等四方面的联系。
\itemize{
\item 输入素材一般放置在IMPORT主题文件夹下，由爬虫应用等从外部写入；
\item 输出结果一般保存在CACHE主题文件夹下，与其他指标构建类似；
\item 处理队列时导入输入素材的处理顺序整理，整理结果保存在QUEUE主题文件下；
\item 处理脚本一般保存在TASK主题文件夹下，按照处理队列中的顺序执行脚本。
}
}
\concept{dataset-import function}
