% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_last_affected}
\alias{ds_last_affected}
\title{读取最近一次更新时重写过分区的数据集文件}
\usage{
ds_last_affected(dsName = c(), stateTopic = "STATE")
}
\arguments{
\item{dsName}{数据集名称}

\item{stateTopic}{状态数据主题域}
}
\description{
读取最近一次更新时重写过分区的数据集文件
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_init}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
