% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_n_max}
\alias{ds_n_max}
\title{按最大值取N条记录}
\usage{
ds_n_max(d, orderColumn, n = 10, with_ties = FALSE)
}
\description{
按最大值取N条记录
}
\seealso{
Other gali-dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_arrange}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()}
}
\concept{gali-dataset function}
