% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_read}
\alias{ds_read}
\title{读取数据集}
\usage{
ds_read(dsName, topic = "CACHE", toFix = TRUE, noDeleted = TRUE)
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}

\item{toFix}{修复因归档后未及时清理__APPEND__数据遗漏的部分}

\item{noDeleted}{不返回标记为删除的数据}
}
\description{
读取数据集
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()}
}
\concept{dataset function}
