% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_read_affected}
\alias{ds_read_affected}
\title{读取重写过分区的数据集文件}
\usage{
ds_read_affected(affectedParts)
}
\arguments{
\item{affectedParts}{受影响的分区文件，多个时用逗号间隔表示}
}
\description{
读取重写过分区的数据集文件
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_last_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
