% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{ds_remove_path}
\alias{ds_remove_path}
\title{移除数据文件夹}
\usage{
ds_remove_path(dsName, topic = "CACHE")
}
\description{
可以整个移除，也可以组装目录后按分区移除
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_schema}()},
\code{\link{last_affected_parts}()},
\code{\link{read_affected_parts}()},
\code{\link{read_dataset}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
