% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{ds_schema}
\alias{ds_schema}
\title{查看数据集架构}
\usage{
ds_schema(ds)
}
\arguments{
\item{ds}{数据框}
}
\description{
查看数据集架构时，必须包含数据
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{last_affected_parts}()},
\code{\link{read_affected_parts}()},
\code{\link{read_dataset}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
