% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_schema_obj}
\alias{ds_schema_obj}
\title{构造Arrow的schema}
\usage{
ds_schema_obj(dsName, topic = "CACHE")
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}
}
\description{
从数据框格式的schema转换为Arrow的Schema对象
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_datetime}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
