% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_select}
\alias{ds_select}
\title{选择列，支持惰性计算}
\usage{
ds_select(d, columns = list(), showOthers = FALSE, regex = NULL)
}
\description{
选择列，支持惰性计算
}
\seealso{
Other gali-dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_arrange}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_tail}()}
}
\concept{gali-dataset function}
