% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_submit}
\alias{ds_submit}
\title{数据集归档}
\usage{
ds_submit(dsName, topic = "CACHE")
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}
}
\description{
消除零散的追加文件
}
\details{
归档操作可能耗时，应当集中处理。
如果不归档，则可能因为零散文件较多，而影响数据读取速度
}
