% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-ds.R
\name{ds_tail}
\alias{ds_tail}
\title{尾部数据}
\usage{
ds_tail(d, n = 10)
}
\description{
尾部数据
}
\seealso{
Other gali-dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_arrange}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_rename}()},
\code{\link{ds_select}()}
}
\concept{gali-dataset function}
