% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{ds_task_queue_init}
\alias{ds_task_queue_init}
\title{初始化队列数据集}
\usage{
ds_task_queue_init(dsName = "__TASK_QUEUE__", cacheTopic = "CACHE")
}
\description{
批处理任务的队列
}
\seealso{
Other task function: 
\code{\link{ds_task_queue_item}()},
\code{\link{ds_task_queue_run}()},
\code{\link{ds_task_queue_todo}()}
}
\concept{task function}
