% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{ds_task_queue_item}
\alias{ds_task_queue_item}
\title{构造队列中的一条数据}
\usage{
ds_task_queue_item(
  taskId,
  params,
  taskType = "__TYPE_UNKNOWN__",
  taskTopic = "CACHE",
  id = gen_batchNum()
)
}
\description{
构造队列中的一条数据
}
\seealso{
Other task function: 
\code{\link{ds_task_queue_init}()},
\code{\link{ds_task_queue_run}()},
\code{\link{ds_task_queue_todo}()}
}
\concept{task function}
