% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{ds_task_queue_run}
\alias{ds_task_queue_run}
\title{批量执行队列任务}
\usage{
ds_task_queue_run(
  batchId,
  runMode = "in-process",
  dsName = "__TASK_QUEUE__",
  cacheTopic = "CACHE"
)
}
\description{
相同批次的任务按优先级大小依次执行；
执行后的结果统一更新到队列。

如果批任务中部分任务失败，则整体批次任务暂停，
批次中的所有任务信息都不更新。
}
\seealso{
Other task function: 
\code{\link{ds_task_queue_init}()},
\code{\link{ds_task_queue_item}()},
\code{\link{ds_task_queue_todo}()}
}
\concept{task function}
