% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_type}
\alias{ds_type}
\title{规范数据存储类型}
\usage{
ds_type(type = "__UNKNOWN__")
}
\description{
规范数据存储类型
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
