% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_write}
\alias{ds_write}
\title{写入数据}
\usage{
ds_write(d, dsName, topic = "CACHE")
}
\arguments{
\item{d}{要追加的数据}

\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}
}
\description{
先执行追加操作，然后直接归档整理。
}
\seealso{
Other dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_arrange}()},
\code{\link{ds_as_datetime}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_delete}()},
\code{\link{ds_demo}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_distinct}()},
\code{\link{ds_drop}()},
\code{\link{ds_dup_count}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_init}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_rename}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()},
\code{\link{ds_type}()}
}
\concept{dataset function}
