% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{ds_write_yaml}
\alias{ds_write_yaml}
\title{写入数据集的Yaml配置}
\usage{
ds_write_yaml(meta, dsName, topic = "CACHE")
}
\arguments{
\item{meta}{元数据内容，支持局部更新}

\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}
}
\description{
允许对配置文件做局部更新
}
\seealso{
Other metadata function: 
\code{\link{ds_yaml}()}
}
\concept{metadata function}
