% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_yaml}
\alias{ds_yaml}
\title{查看数据集元数据}
\usage{
ds_yaml(dsName, topic = "CACHE")
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}
}
\description{
查看数据集元数据
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_last_affected}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
