% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-yaml.R
\name{ds_yaml_schema}
\alias{ds_yaml_schema}
\title{将yaml配置中的架构转化为数据框}
\usage{
ds_yaml_schema(dsName, topic = "CACHE")
}
\description{
将yaml配置中的架构转化为数据框
}
\seealso{
Other metadata function: 
\code{\link{ds_yaml_write}()},
\code{\link{ds_yaml}()}
}
\concept{metadata function}
