% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-DT.R
\name{DT_table}
\alias{DT_table}
\title{数据表}
\usage{
DT_table(
  d,
  rownames = FALSE,
  selection = "none",
  extensions = "Responsive",
  class = "white-space: nowrap",
  options = list(responsive = TRUE),
  ...
)
}
\description{
数据表
}
\examples{
mtcars |> DT_table()
}
\concept{lib-DT functions}
