% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form_read}
\alias{form_read}
\title{读取表单数据}
\usage{
form_read(formId, formTopic = "FORM")
}
\arguments{
\item{formId}{每个批次号保存为一个文件}

\item{formTopic}{保存表单的存储主题文件夹}

\item{family}{form function}
}
\description{
读取队列中待执行的任务的输入参数
}
