% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form_write}
\alias{form_write}
\title{写入表单数据}
\usage{
form_write(data, formId, formTopic = "FORM")
}
\arguments{
\item{data}{表单数据，一般以列表结构保存}

\item{formId}{保障唯一，每个\code{formId}会保存为一个独立文件}

\item{formTopic}{保存表单的存储主题文件夹}

\item{family}{form function}
}
\description{
为队列中待执行的任务保存输入参数
}
