% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{gali_arrange}
\alias{gali_arrange}
\title{行排序}
\usage{
gali_arrange(
  sv_columns = list(),
  b_desc = FALSE,
  b_by_group = FALSE,
  s_dataName = "@result"
)
}
\description{
行排序
}
\seealso{
Other data-plyr function: 
\code{\link{gali_collect}()},
\code{\link{gali_filter}()},
\code{\link{gali_head}()},
\code{\link{gali_n_max}()},
\code{\link{gali_n_min}()},
\code{\link{gali_read}()},
\code{\link{gali_rename}()},
\code{\link{gali_select}()},
\code{\link{gali_tail}()}
}
\concept{data-plyr function}
