% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{gali_create_risk_model}
\alias{gali_create_risk_model}
\title{定义风险模型}
\usage{
gali_create_risk_model(
  s_dsName,
  s_modelName,
  s_tagName = "main",
  s_riskTip = "-",
  e_riskLevel = "L",
  s_modelDesc = "-",
  s_author = "-"
)
}
\description{
定义风险模型
}
\seealso{
Other gali-risk function: 
\code{\link{gali_read_risk}()},
\code{\link{gali_search_risk_model}()},
\code{\link{gali_write_risk}()}
}
\concept{gali-risk function}
