% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_dataset_arrange}
\alias{gali_dataset_arrange}
\title{行排序}
\usage{
gali_dataset_arrange(
  sv_columns = list(),
  b_desc = FALSE,
  b_by_group = FALSE,
  s_dataName = "@result"
)
}
\description{
行排序
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_collect}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_head}()},
\code{\link{gali_dataset_n_max}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_rename}()},
\code{\link{gali_dataset_select}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
