% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_dataset_collect}
\alias{gali_dataset_collect}
\title{立即执行数据收集（结束惰性计算）}
\usage{
gali_dataset_collect(s_dataName = "@result")
}
\description{
立即执行数据收集（结束惰性计算）
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_arrange}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_head}()},
\code{\link{gali_dataset_n_max}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_rename}()},
\code{\link{gali_dataset_select}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_read}()}
}
\concept{gali-dataset function}
