% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_dataset_head}
\alias{gali_dataset_head}
\title{头部数据}
\usage{
gali_dataset_head(i_n = 10, s_dataName = "@result")
}
\description{
头部数据
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_arrange}()},
\code{\link{gali_dataset_collect}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_n_max}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_rename}()},
\code{\link{gali_dataset_select}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_read_dataset}()}
}
\concept{gali-dataset function}
