% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_dataset_n_max}
\alias{gali_dataset_n_max}
\title{按最大值取N条记录}
\usage{
gali_dataset_n_max(
  s_orderColumn,
  i_n = 10,
  b_with_ties = FALSE,
  s_dataName = "@result"
)
}
\description{
按最大值取N条记录
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_arrange}()},
\code{\link{gali_dataset_collect}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_head}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_rename}()},
\code{\link{gali_dataset_select}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
