% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_dataset_rename}
\alias{gali_dataset_rename}
\title{列改名}
\usage{
gali_dataset_rename(s_newName, s_oldName, s_dataName = "@result")
}
\description{
列改名
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_arrange}()},
\code{\link{gali_dataset_collect}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_head}()},
\code{\link{gali_dataset_n_max}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_select}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
