% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_dataset_select}
\alias{gali_dataset_select}
\title{选择列，支持惰性计算}
\usage{
gali_dataset_select(
  sv_columns = list(),
  b_everything = FALSE,
  s_regex = NULL,
  s_dataName = "@result"
)
}
\description{
选择列，支持惰性计算
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_arrange}()},
\code{\link{gali_dataset_collect}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_head}()},
\code{\link{gali_dataset_n_max}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_rename}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
