% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_ds_arrange}
\alias{gali_ds_arrange}
\title{行排序}
\usage{
gali_ds_arrange(
  d = NULL,
  sv_columns = list(),
  b_desc = FALSE,
  b_by_group = FALSE
)
}
\description{
行排序
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_ds_collect}()},
\code{\link{gali_ds_filter}()},
\code{\link{gali_ds_head}()},
\code{\link{gali_ds_n_max}()},
\code{\link{gali_ds_n_min}()},
\code{\link{gali_ds_rename}()},
\code{\link{gali_ds_select}()},
\code{\link{gali_ds_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
