% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_ds_collect}
\alias{gali_ds_collect}
\title{立即执行数据收集（结束惰性计算）}
\usage{
gali_ds_collect(d = NULL)
}
\description{
立即执行数据收集（结束惰性计算）
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_ds_arrange}()},
\code{\link{gali_ds_filter}()},
\code{\link{gali_ds_head}()},
\code{\link{gali_ds_n_max}()},
\code{\link{gali_ds_n_min}()},
\code{\link{gali_ds_rename}()},
\code{\link{gali_ds_select}()},
\code{\link{gali_ds_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
