% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_ds_filter}
\alias{gali_ds_filter}
\title{定义数据过滤任务}
\usage{
gali_ds_filter(d = NULL, s_column, o_name, fv_value = list(NULL))
}
\arguments{
\item{s_column}{列名}

\item{o_name}{阈值范围判断符号}

\item{fv_value}{阈值}

\item{taskId}{任务ID}

\item{s_dataName}{内存中的数据框名称，默认为 @result}

\item{taskTopic}{任务定义的主题文件夹}
}
\description{
这个任务的目标是通过配置项实现函数\code{\link[dplyr:filter]{dplyr::filter()}}的大部分功能。

允许为数据集增加多个阈值查询条件，缩小筛查范围。

taskId、s_column、o_name、fv_value等参数构造唯一的gali_ds_filter表达式，
这将允许从UI生成或还原该操作。

只要包括以下逻辑判断：
\itemize{
\item >, <, >=, <=, ==, !=
\item \%in\%, \%nin\%
\item \%regex\%, \%not-regex\%
\item #time# >, #time# <
\item #date# >, #date# <
}
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_ds_add_count}()},
\code{\link{gali_ds_arrange}()},
\code{\link{gali_ds_collect}()},
\code{\link{gali_ds_count}()},
\code{\link{gali_ds_head}()},
\code{\link{gali_ds_n_max}()},
\code{\link{gali_ds_n_min}()},
\code{\link{gali_ds_rename}()},
\code{\link{gali_ds_select}()},
\code{\link{gali_ds_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
