% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_ds_n_max}
\alias{gali_ds_n_max}
\title{按最大值取N条记录}
\usage{
gali_ds_n_max(d = NULL, s_orderColumn, i_n = 10, b_with_ties = FALSE)
}
\description{
按最大值取N条记录
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_ds_arrange}()},
\code{\link{gali_ds_collect}()},
\code{\link{gali_ds_filter}()},
\code{\link{gali_ds_head}()},
\code{\link{gali_ds_n_min}()},
\code{\link{gali_ds_rename}()},
\code{\link{gali_ds_select}()},
\code{\link{gali_ds_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
