% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_ds_select}
\alias{gali_ds_select}
\title{选择列，支持惰性计算}
\usage{
gali_ds_select(
  `@ds`,
  sv_columns = list(),
  b_everything = FALSE,
  s_regex = NULL
)
}
\description{
选择列，支持惰性计算
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_ds_add_count}()},
\code{\link{gali_ds_arrange}()},
\code{\link{gali_ds_collect}()},
\code{\link{gali_ds_count}()},
\code{\link{gali_ds_filter}()},
\code{\link{gali_ds_head}()},
\code{\link{gali_ds_n_max}()},
\code{\link{gali_ds_n_min}()},
\code{\link{gali_ds_rename}()},
\code{\link{gali_ds_tail}()},
\code{\link{gali_read}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
