% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-DT.R
\name{gali_DT}
\alias{gali_DT}
\title{数据表}
\usage{
gali_DT(
  d,
  rownames = FALSE,
  selection = "none",
  extensions = "Responsive",
  class = "white-space: nowrap",
  options = list(responsive = TRUE),
  ...
)
}
\description{
数据表
}
\examples{
mtcars |> gali_DT()
}
\seealso{
Other task-lib functions: 
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{task-lib functions}
