% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{gali_n_min}
\alias{gali_n_min}
\title{按最小值取N条记录}
\usage{
gali_n_min(
  s_orderColumn,
  i_n = 10,
  b_with_ties = FALSE,
  s_dataName = "@result"
)
}
\description{
按最小值取N条记录
}
\seealso{
Other data-plyr function: 
\code{\link{gali_arrange}()},
\code{\link{gali_collect}()},
\code{\link{gali_filter}()},
\code{\link{gali_head}()},
\code{\link{gali_n_max}()},
\code{\link{gali_read}()},
\code{\link{gali_rename}()},
\code{\link{gali_select}()},
\code{\link{gali_tail}()}
}
\concept{data-plyr function}
