% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{gali_plot_area}
\alias{gali_plot_area}
\title{面积图}
\usage{
gali_plot_area(
  `@ds`,
  e_x,
  e_y,
  c_fill = "tozeroy",
  c_color = "red",
  e_shape = "spline"
)
}
\description{
面积图
}
\examples{
mtcars |>
  count(cyl) |>
  plot_area(x = "cyl", y = "n")
  
mtcars |> group_by(cyl) |>
  summarise(displ = mean(disp)) |>
  gali_plot_area(e_x = "cyl", e_y = "displ")
  
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie2}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
