% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-plotly.R
\name{gali_plot_bar}
\alias{gali_plot_bar}
\title{柱图}
\usage{
gali_plot_bar(`@ds`, e_x, e_y)
}
\description{
柱图
}
\examples{
mtcars |>
  count(cyl) |>
  gali_plot_bar(e_x = "cyl", e_y = "n")

mpg |> count(class) |>
  mutate(class = forcats::fct_reorder(class, n, .desc = TRUE)) |>
  plot_bar(x = "class", y = "n")
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie2}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
