% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-plotly.R
\name{gali_plot_barpolar}
\alias{gali_plot_barpolar}
\title{极坐标柱}
\usage{
gali_plot_barpolar(
  `@ds`,
  e_theta,
  e_r,
  c_color = NULL,
  i_bargap = 0,
  f1_hole = 0.05,
  e_direction = "clockwise"
)
}
\description{
极坐标柱
}
\examples{
mtcars |>
  count(cyl) |>
  mutate(cyl = sprintf("CYL: \%d", cyl)) |>
  gali_plot_barpolar(e_theta = "cyl", e_r = "n")
  
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie2}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
