% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{gali_plot_hist}
\alias{gali_plot_hist}
\title{直方图}
\usage{
gali_plot_hist(`@ds`, e_x, c_color = "brown", s_xtitle = NULL, s_ytitle = NULL)
}
\description{
直方图
}
\examples{
mtcars |> gali_plot_hist(e_x = "mpg")
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie2}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
