% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{gali_plot_line}
\alias{gali_plot_line}
\title{线图}
\usage{
gali_plot_line(
  `@ds`,
  e_x,
  e_y,
  e_fillcolor = "red",
  c_linecolor = "rgb(205, 12, 24)",
  e_shape = "spline"
)
}
\description{
线图
}
\examples{
mtcars |>
  count(cyl) |> 
  gali_plot_line(e_x = "cyl", e_y = "n")

}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie2}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
