% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{gali_plot_marker}
\alias{gali_plot_marker}
\title{散点图}
\usage{
gali_plot_marker(`@ds`, e_x, e_y, f1_alpha = 0.2, s_name = NULL)
}
\description{
散点图
}
\examples{
mtcars |>
 plot_marker(e_x = "mpg", e_y = "disp", f1_alpha = 0.6)
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_pie2}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
