% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-plotly.R
\name{gali_plot_pie}
\alias{gali_plot_pie}
\title{饼图}
\usage{
gali_plot_pie(`@ds`, e_value, s_label = NULL, f1_hole = 0.4)
}
\description{
饼图
}
\examples{
mtcars |> gali_plot_pie(e_value = "cyl")
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie2}()}
}
\concept{lib-plotly functions}
