% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-plotly.R
\name{gali_plot_pie2}
\alias{gali_plot_pie2}
\title{饼图}
\usage{
gali_plot_pie2(`@ds`, e_value, s_label = NULL, f1s_pull = 0, f1_hole = 0.25)
}
\description{
饼图
}
\examples{
mtcars |> count(cyl) |>
 gali_plot_pie2(e_value = "n", pull = c(0.1, 0, 0))
}
\seealso{
Other lib-plotly functions: 
\code{\link{gali_plot_area}()},
\code{\link{gali_plot_barpolar}()},
\code{\link{gali_plot_bar}()},
\code{\link{gali_plot_hist}()},
\code{\link{gali_plot_line}()},
\code{\link{gali_plot_marker}()},
\code{\link{gali_plot_pie}()}
}
\concept{lib-plotly functions}
