% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{gali_read}
\alias{gali_read}
\title{读取数据集}
\usage{
gali_read(s_dsName, s_dataName = "@result")
}
\description{
读取数据集
}
\seealso{
Other data-plyr function: 
\code{\link{gali_arrange}()},
\code{\link{gali_collect}()},
\code{\link{gali_filter}()},
\code{\link{gali_head}()},
\code{\link{gali_n_max}()},
\code{\link{gali_n_min}()},
\code{\link{gali_rename}()},
\code{\link{gali_select}()},
\code{\link{gali_tail}()}
}
\concept{data-plyr function}
