% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{gali_read_risk}
\alias{gali_read_risk}
\title{读取疑点数据}
\usage{
gali_read_risk(b_todoFlag = TRUE)
}
\description{
读取疑点数据
}
\seealso{
Other gali-risk function: 
\code{\link{gali_create_risk_model}()},
\code{\link{gali_search_risk_model}()},
\code{\link{gali_write_risk}()}
}
\concept{gali-risk function}
