% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{gali_rename}
\alias{gali_rename}
\title{列改名}
\usage{
gali_rename(s_newName, s_oldName, s_dataName = "@result")
}
\description{
列改名
}
\seealso{
Other data-plyr function: 
\code{\link{gali_arrange}()},
\code{\link{gali_collect}()},
\code{\link{gali_filter}()},
\code{\link{gali_head}()},
\code{\link{gali_n_max}()},
\code{\link{gali_n_min}()},
\code{\link{gali_read}()},
\code{\link{gali_select}()},
\code{\link{gali_tail}()}
}
\concept{data-plyr function}
