% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{gali_search_risk_model}
\alias{gali_search_risk_model}
\title{查找风险模型}
\usage{
gali_search_risk_model(s_modelMatch = ".*")
}
\description{
查找风险模型
}
\seealso{
Other gali-risk function: 
\code{\link{gali_create_risk_model}()},
\code{\link{gali_read_risk}()},
\code{\link{gali_write_risk}()}
}
\concept{gali-risk function}
