% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{gali_select}
\alias{gali_select}
\title{选择列，支持惰性计算}
\usage{
gali_select(
  sv_columns = list(),
  b_everything = FALSE,
  s_regex = NULL,
  s_dataName = "@result"
)
}
\description{
选择列，支持惰性计算
}
\seealso{
Other data-plyr function: 
\code{\link{gali_arrange}()},
\code{\link{gali_collect}()},
\code{\link{gali_filter}()},
\code{\link{gali_head}()},
\code{\link{gali_n_max}()},
\code{\link{gali_n_min}()},
\code{\link{gali_read}()},
\code{\link{gali_rename}()},
\code{\link{gali_tail}()}
}
\concept{data-plyr function}
