% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{gali_tail}
\alias{gali_tail}
\title{尾部数据}
\usage{
gali_tail(i_n = 10, s_dataName = "@result")
}
\description{
尾部数据
}
\seealso{
Other data-plyr function: 
\code{\link{gali_arrange}()},
\code{\link{gali_collect}()},
\code{\link{gali_filter}()},
\code{\link{gali_head}()},
\code{\link{gali_n_max}()},
\code{\link{gali_n_min}()},
\code{\link{gali_read}()},
\code{\link{gali_rename}()},
\code{\link{gali_select}()}
}
\concept{data-plyr function}
