% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_write}
\alias{gali_write}
\title{写入数据集}
\usage{
gali_write(`@ds`, e_dsName)
}
\description{
写入数据集
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_ds_add_count}()},
\code{\link{gali_ds_arrange}()},
\code{\link{gali_ds_collect}()},
\code{\link{gali_ds_count}()},
\code{\link{gali_ds_filter}()},
\code{\link{gali_ds_head}()},
\code{\link{gali_ds_n_max}()},
\code{\link{gali_ds_n_min}()},
\code{\link{gali_ds_rename}()},
\code{\link{gali_ds_select}()},
\code{\link{gali_ds_tail}()},
\code{\link{gali_read}()}
}
\concept{gali-dataset function}
