% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{gali_write_risk}
\alias{gali_write_risk}
\title{根据风险模型生成并写入疑点数据}
\usage{
gali_write_risk(d = NULL, s_modelId)
}
\description{
根据风险模型生成并写入疑点数据
}
\seealso{
Other gali-risk function: 
\code{\link{gali_create_risk_model}()},
\code{\link{gali_read_risk}()},
\code{\link{gali_search_risk_model}()}
}
\concept{gali-risk function}
