% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_batchNum}
\alias{gen_batchNum}
\title{获取批次号}
\usage{
gen_batchNum()
}
\description{
获取批次号
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%error\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
