% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{读取所有配置项}
\usage{
get_config()
}
\description{
读取所有配置项
}
\seealso{
Other config functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{load_config}()},
\code{\link{set_topic}()}
}
\concept{config functions}
