% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{get_importing_folder}
\alias{get_importing_folder}
\title{查看当前导入文件夹}
\usage{
get_importing_folder()
}
\description{
查看当前导入文件夹
}
\seealso{
Other task functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{load_config}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
