% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_params}
\alias{get_params}
\title{获取函数参数}
\usage{
get_params(funcName)
}
\arguments{
\item{funcName}{函数名称，可以是字符串或函数名}
}
\description{
根据规则生成参数要求

如果参数命名格式按照如下规范约定，则自动生成参数要求
\itemize{
\item s_ 要求单个字符串
\item sn_ 要求多个字符串
\item i_ 要求整数
\item f_ 要求浮点数
\item ds_ 要求日期格式字符串
\item dts_ 要求日期时间格式字符串
\item ts_ 要求时间格式字符串
\item t_ 要求时间戳整数
\item c_ 颜色枚举
\item b_ 要求布尔类型
\item e_ 要求枚举值，要求在配置中额外定义
可以是静态文件定义，也可以是从某个数据集中实时提取
}
}
