% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{get_params_plot}
\alias{get_params_plot}
\title{查询}
\usage{
get_params_plot(
  entry = "plotly_schema",
  paramName = NULL,
  v = NULL,
  plot = NULL
)
}
\arguments{
\item{paramName}{参数名称}

\item{plot}{plotly对象}

\item{type}{trace类型}
}
\description{
查询
}
