% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{get_path}
\alias{get_path}
\title{根据配置构造数据路径}
\usage{
get_path(topic, ...)
}
\arguments{
\item{topic}{任务主题}
}
\description{
根据配置构造数据路径
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{load_config}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
