% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-folder.R
\name{get_task_plan}
\alias{get_task_plan}
\title{查看执行计划}
\usage{
get_task_plan(path, glob = "*.R")
}
\arguments{
\item{path}{文件夹位置}

\item{glob}{要执行的源文件默认以.R结尾}
}
\description{
按执行顺序罗列需要执行的脚本
}
\seealso{
Other TaskFolder functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{run_task}()},
\code{\link{set_topic}()}
}
\concept{TaskFolder functions}
