% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimmer-package.R
\docType{package}
\name{glimmer-package}
\alias{glimmer}
\alias{glimmer-package}
\title{glimmer: A database for R Scripts based on Apache Arrow}
\description{
Help R uses to write less code and do more.
}
\author{
\strong{Maintainer}: homeway Xue \email{xuehongwei@qq.com}

}
\keyword{internal}
