% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%empty\%}
\alias{\%empty\%}
\title{遇到空值填充默认值}
\usage{
a \%empty\% b
}
\description{
当值为NULL、空向量、空列表、空字符串时设置默认值
}
\seealso{
Other utils function: 
\code{\link{\%error\%}()},
\code{\link{\%na\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{\%not-na\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{gen_batchNum}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
