% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%error\%}
\alias{\%error\%}
\title{遇到异常填充默认值}
\usage{
exp \%error\% defaultValue
}
\description{
遇到异常填充默认值
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%na\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{\%not-na\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{gen_batchNum}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
