% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not-empty\%}
\alias{\%not-empty\%}
\title{遇到非空值填充默认值}
\usage{
a \%not-empty\% b
}
\description{
遇到非空值填充默认值
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%error\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{gen_batchNum}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
