% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not-regex\%}
\alias{\%not-regex\%}
\title{条件过滤：使用正则表达式取反}
\usage{
a \%not-regex\% b
}
\description{
条件过滤：使用正则表达式取反
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%error\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{gen_batchNum}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
