% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%regex\%}
\alias{\%regex\%}
\title{条件过滤：使用正则表达式}
\usage{
a \%regex\% b
}
\description{
条件过滤：使用正则表达式
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%error\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{create_dir}()},
\code{\link{gen_batchNum}()},
\code{\link{remove_dir}()}
}
\concept{utils function}
