% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_changed}
\alias{import_changed}
\title{检测发生变化的素材文件}
\usage{
import_changed(importTopic = "IMPORT", snapTopic = "SNAP")
}
\description{
检测发生变化的素材文件
}
\seealso{
Other import function: 
\code{\link{import_csv_preview}()},
\code{\link{import_csv}()},
\code{\link{import_files_preview}()},
\code{\link{import_files}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_task_create}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
