% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_csv_files}
\alias{import_csv_files}
\title{导入csv文件}
\usage{
import_csv_files(dsName, keyColumns, filesMatched, importTopic)
}
\arguments{
\item{filesMatched}{导入素材文件名}

\item{importTopic}{IMPORT}
}
\description{
导入csv文件
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_init}()},
\code{\link{import_run}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_task_create}()}
}
\concept{import function}
