% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_csv_preview}
\alias{import_csv_preview}
\title{导入csv文件}
\usage{
import_csv_preview(filesMatched, func = NULL)
}
\arguments{
\item{filesMatched}{导入素材文件名}
}
\description{
导入csv文件
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_csv}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_task_create}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
