% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dataset_init}
\alias{import_dataset_init}
\title{初始化导入状态数据集}
\usage{
import_dataset_init(dsName = "__IMPORT_FILES__", cacheTopic = "CACHE")
}
\description{
导入素材中所有文件都将入库。

每次写入的导入素材文件，应当放置在同一个批次文件夹中。
该批次文件夹应当包含时间要素，以方便按写入时间排序。

在同一次扫描中，如果不同批次文件夹中的有相同路径的导入素材文件，只关注最新文件
如果有补充采集，则视情况将旧文件标记为“忽略”。
}
\seealso{
Other import function: 
\code{\link{ds_import_match}()},
\code{\link{ds_import_write}()},
\code{\link{import_files_all}()},
\code{\link{import_task_create}()},
\code{\link{import_task_run}()}
}
\concept{import function}
