% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dataset_task_match}
\alias{import_dataset_task_match}
\title{为导入素材匹配任务}
\usage{
import_dataset_task_match(
  importDataset = "__IMPORT_FILES__",
  taskTopic = "TASK_DEFINE",
  cacheTopic = "CACHE",
  onlyNewFiles = TRUE
)
}
\description{
开发过程中，可能要为新建任务重新匹配导入素材。

此时，可以根据实际情况决定仅针对未处理状态还是所有的导入素材做匹配。
}
\seealso{
Other import function: 
\code{\link{import_dataset_read}()},
\code{\link{import_files_all}()},
\code{\link{import_files_scan}()},
\code{\link{import_init}()},
\code{\link{import_task_queue_create}()},
\code{\link{import_task_queue_run}()}
}
\concept{import function}
