% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{import_define}
\alias{import_define}
\title{定义从任务文件夹导入数据集的函数}
\usage{
import_define(dsName, fun, topic = "IMPORT")
}
\arguments{
\item{dsName}{要导入的数据集名称}

\item{fun}{导入函数定义，要求是一个匿名函数}

\item{topic}{导入文件夹所在的主题域}
}
\description{
在定义增量导入任务时需要定义按数据集导入的函数
}
\details{
所定义的导入函数，需要读取数据，就要组装数据所在位置，规则如下：
{IMPORT}/{\strong{DOING_TASK_FOLDER}}/{dsName}
其中，__DOING_TASK_FOLDER__就是当前正在处理的任务文件夹名称，
而dsName就是任务文件夹下的数据集文件夹，数据应当放在dsName文件夹内

举一个例子，如果IMPORT目录位置是："~/glimmer/IMPORT"
要导入的数据位置是："~/glimmer/IMPORT/task1/mycsv/1.csv"
其中，task1就是导入文件夹，mycsv就是数据集。

考虑增量导入的情况，在新的数据到位后，
可能多了一个这样的文件："~/glimmer/IMPORT/task2/mycsv/1.csv"
此时要做的，就是针对task2进行导入。
而glimmer包提供的import_todo函数就是自动识别出新增的task2，并执行导入脚本。
}
\concept{dataset functions}
