% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_files}
\alias{import_files}
\title{批量导入文件}
\usage{
import_files(filesMatched, dsName, keyColumns = c(), func = function(path) {
})
}
\arguments{
\item{filesMatched}{导入素材文件名}

\item{dsName}{导出目标数据集名称}

\item{keyColumns}{数据集的关键字段}
}
\description{
批量导入文件
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_csv_preview}()},
\code{\link{import_csv}()},
\code{\link{import_files_preview}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_task_create}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
