% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_files_preview}
\alias{import_files_preview}
\title{预览多个文件}
\usage{
import_files_preview(filesMatched, func = function(path) {
})
}
\arguments{
\item{filesMatched}{导入素材文件名}
}
\description{
预览多个文件
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_csv_preview}()},
\code{\link{import_csv}()},
\code{\link{import_files}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_task_create}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
