% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{import_folders}
\alias{import_folders}
\title{所有导入文件夹}
\usage{
import_folders(importTopic = "IMPORT", taskTopic = "TASK/IMPORT")
}
\arguments{
\item{importTopic}{导入主题域，即导入素材的根目录}
}
\description{
将主题域下的所有一级子目录作为导入素材
}
\details{
未导入成功过的文件夹可作为增量导入素材，被自动发现。
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
