% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_folders_all}
\alias{import_folders_all}
\title{列举所有文件}
\usage{
import_folders_all(importTopic = "IMPORT")
}
\description{
列举所有文件
}
\seealso{
Other import function: 
\code{\link{import_dataset_init}()},
\code{\link{import_folders_ignore}()},
\code{\link{import_folders_scan}()},
\code{\link{import_task_gen}()},
\code{\link{import_task_run}()}
}
\concept{import function}
