% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_folders_ignore}
\alias{import_folders_ignore}
\title{将导入素材标记为忽略状态}
\usage{
import_folders_ignore()
}
\description{
将导入素材标记为忽略状态
}
\seealso{
Other import function: 
\code{\link{import_dataset_init}()},
\code{\link{import_folders_all}()},
\code{\link{import_folders_scan}()},
\code{\link{import_task_gen}()},
\code{\link{import_task_run}()}
}
\concept{import function}
