% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_folders_scan}
\alias{import_folders_scan}
\title{从导入素材中提取未扫描文件}
\usage{
import_folders_scan(
  importDataset = "__IMPORT_FILES__",
  importTopic = "IMPORT",
  cacheTopic = "CACHE"
)
}
\description{
读取导入素材，同时读取导入状态库，
将未曾导入或有修改时间有变动的文件写入状态库。

处理导入任务时也将修改这一状态库，标记taskReadAt、taskDoneAt、taskDoneAt、ignore等字段。
}
\seealso{
Other import function: 
\code{\link{import_dataset_init}()},
\code{\link{import_folders_all}()},
\code{\link{import_folders_ignore}()},
\code{\link{import_task_gen}()},
\code{\link{import_task_run}()}
}
\concept{import function}
