% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_run}
\alias{import_run}
\title{执行所有导入任务}
\usage{
import_run(
  files = tibble(),
  tasks = tibble(),
  toImport = TRUE,
  taskTopic = "TASK_DEFINE",
  importTopic = "IMPORT",
  importDataset = "__IMPORT_FILES__",
  cacheTopic = "CACHE"
)
}
\description{
对于新素材，或指定素材文件，执行所有任务检查。
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_task_create}()}
}
\concept{import function}
