% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_search}
\alias{import_search}
\title{扫描所有未处理、未忽略的文件}
\usage{
import_search(
  fileMatch = ".*",
  batchMatch = ".*",
  importDataset = "__IMPORT_FILES__",
  cacheTopic = "CACHE",
  ignoreFlag = FALSE,
  b_todo = TRUE
)
}
\description{
扫描所有未处理、未忽略的文件
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_init}()},
\code{\link{import_run}()},
\code{\link{import_scan}()}
}
\concept{import function}
