% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_task_create}
\alias{import_task_create}
\title{定义导入任务}
\usage{
import_task_create(
  taskId,
  importFileMatch,
  targetDataset,
  importTopic = "IMPORT",
  cacheTopic = "CACHE",
  taskTopic = "TASK_DEFINE",
  scriptsTopic = "TASK_SCRIPTS"
)
}
\description{
定义导入任务
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_csv_files}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()},
\code{\link{import_todo_flag}()}
}
\concept{import function}
