% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_task_gen}
\alias{import_task_gen}
\title{生成导入任务}
\usage{
import_task_gen(
  taskDataset = "__TASK_QUEUE__",
  importDataset = "__IMPORT_FILES__",
  taskTopic = "TASK_DEFINE",
  importTopic = "IMPORT",
  cacheTopic = "CACHE"
)
}
\description{
生成导入任务
}
\seealso{
Other import function: 
\code{\link{import_dataset_init}()},
\code{\link{import_folders_all}()},
\code{\link{import_folders_ignore}()},
\code{\link{import_folders_scan}()},
\code{\link{import_task_run}()}
}
\concept{import function}
