% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_task_queue_create}
\alias{import_task_queue_create}
\title{构建导入任务的批处理队列}
\usage{
import_task_queue_create(
  taskQueue = "__TASK_QUEUE__",
  importDataset = "__IMPORT_FILES__",
  taskTopic = "TASK_DEFINE",
  importTopic = "IMPORT",
  cacheTopic = "CACHE"
)
}
\description{
扫描素材时，要求遵循文件夹命名约定：即被扫描文件相对于批次文件夹的相对路径名，
以扫描任务TaskId开始，例如\code{联系人/重要/1.csv}可以匹配TaskID为\code{联系人}的任务，
也可以同时匹配到\code{联系人/重要}的任务。
匹配到的任务都会将素材文件作为参数执行。

扫描后的任务处理流程如下：

（1）将未处理的重复素材设置忽略标记；

（2）罗列所有扫描任务；

（3）根据扫描任务优先级，查找未处理素材，添加导入任务；

（4）导入素材添加后，更新这些素材文件的状态信息（读取时间）。

（5）每次处理的素材文件个数不超过100个，超过时建立新任务；

（6）追加一个任务：导入结束后，更新这些素材文件的状态信息。
}
\seealso{
Other import function: 
\code{\link{import_dataset_read}()},
\code{\link{import_dataset_task_match}()},
\code{\link{import_files_all}()},
\code{\link{import_files_scan}()},
\code{\link{import_init}()},
\code{\link{import_task_queue_run}()}
}
\concept{import function}
