% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_task_run}
\alias{import_task_run}
\title{执行导入任务}
\usage{
import_task_run()
}
\description{
执行导入任务
}
\seealso{
Other import function: 
\code{\link{import_dataset_init}()},
\code{\link{import_folders_all}()},
\code{\link{import_folders_ignore}()},
\code{\link{import_folders_scan}()},
\code{\link{import_task_gen}()}
}
\concept{import function}
