% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{import_todo}
\alias{import_todo}
\title{未曾处理过的任务文件夹}
\usage{
import_todo(importTopic = "IMPORT", taskTopic = "TASK/IMPORT")
}
\description{
从状态库中比对，哪些任务尚未处理过，然后自动执行
}
\details{
脚本主要有两类，一是需要任务文件夹输入的，二是不需要任务文件夹的

第一种情况，典型的是定时导入任务，这很可能是多个任务文件夹

首先，要循环检查所有任务文件夹，
要针对每一个任务文件夹执行脚本，
直到循环完所有任务文件夹为止。

其次，脚本任务也可能包含很多子任务，
需要针对每个任务文件夹，按顺序执行所有任务。
这些任务可能是各自逐项检查任务文件夹中的每个子文件夹是否需要处理。

第二种情况，典型的是风险指标加工，这可能不需要指定任何任务文件夹，
而仅仅是根据已经更新的数据集，做相对复杂的分析和预测，形成疑点数据
}
\seealso{
Other task functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_redo}()},
\code{\link{set_topic}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
