% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_todo_flag}
\alias{import_todo_flag}
\title{修改todo标记}
\usage{
import_todo_flag(
  filesMatched,
  todoFlag = FALSE,
  cacheTopic = "CACHE",
  importDataset = "__IMPORT_FILES__"
)
}
\arguments{
\item{filesMatched}{导入素材文件名}

\item{cacheTopic}{数据集存储主题文件夹}

\item{importDataset}{导入素材库所在的数据集}
}
\description{
修改todo标记
}
\seealso{
Other import function: 
\code{\link{import_changed}()},
\code{\link{import_csv_preview}()},
\code{\link{import_csv}()},
\code{\link{import_files_preview}()},
\code{\link{import_files}()},
\code{\link{import_init}()},
\code{\link{import_scan}()},
\code{\link{import_search}()}
}
\concept{import function}
