% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{load_config}
\alias{load_config}
\title{加载配置文件}
\usage{
load_config(yml_file)
}
\description{
加载配置文件
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
