% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_barpolar}
\alias{plot_barpolar}
\title{极坐标柱}
\usage{
plot_barpolar(
  d,
  e_theta,
  e_r,
  c_color = NULL,
  i_bargap = 0,
  f1_hole = 0.05,
  e_direction = "clockwise"
)
}
\description{
极坐标柱
}
\examples{
mtcars |>
  count(cyl) |>
  mutate(cyl = sprintf("CYL: \%d", cyl)) |>
  gali_plot_barpolar(e_theta = "cyl", e_r = "n")
  
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot0}()},
\code{\link{plot_area}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
