% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{plot_hist}
\alias{plot_hist}
\title{直方图}
\usage{
plot_hist(d, x, color = "brown", xtitle = NULL, ytitle = NULL)
}
\description{
直方图
}
\examples{
mtcars |> plot_hist(x = "mpg")
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{lib-plotly functions}
