% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_marker}
\alias{plot_marker}
\title{散点图}
\usage{
plot_marker(d, e_x, e_y, f1_alpha = 0.2, s_name = NULL)
}
\description{
散点图
}
\examples{
mtcars |>
 plot_marker(e_x = "mpg", e_y = "disp", f1_alpha = 0.6)
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_init}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
