% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{plot_pie2}
\alias{plot_pie2}
\title{饼图}
\usage{
plot_pie2(d, value, label = NULL, pull = 0, hole = 0.25)
}
\description{
饼图
}
\examples{
mtcars |> count(cyl) |>
 plot_pie2(value = "n", pull = c(0.1, 0, 0))
}
\seealso{
Other task-lib functions: 
\code{\link{gali_DT}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie}()}
}
\concept{task-lib functions}
