% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_dataset_init}
\alias{queue_dataset_init}
\title{初始化队列数据集}
\usage{
queue_dataset_init(dsName = "__TASK_QUEUE__", cacheTopic = "CACHE")
}
\description{
批处理任务的队列
}
\seealso{
Other task function: 
\code{\link{queue_run}()}
}
\concept{task function}
