% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_run}
\alias{queue_run}
\title{批量执行队列任务}
\usage{
queue_run(dsName = "__TASK_QUEUE__", cacheTopic = "CACHE")
}
\description{
批量执行队列任务
}
\seealso{
Other task function: 
\code{\link{queue_dataset_init}()}
}
\concept{task function}
