% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_task_item}
\alias{queue_task_item}
\title{构造队列中的一条数据}
\usage{
queue_task_item(
  taskId,
  params,
  taskType = "__TYPE_UNKNOWN__",
  taskTopic = "CACHE"
)
}
\description{
构造队列中的一条数据
}
\seealso{
Other task function: 
\code{\link{queue_batch_run}()},
\code{\link{queue_batch_todo}()},
\code{\link{queue_dataset_init}()}
}
\concept{task function}
