% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{read_affected_parts}
\alias{read_affected_parts}
\title{读取重写过分区的数据集文件}
\usage{
read_affected_parts(affectedParts)
}
\description{
读取重写过分区的数据集文件
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{last_affected_parts}()},
\code{\link{read_dataset}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
