% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{read_dataset}
\alias{read_dataset}
\title{读取数据集}
\usage{
read_dataset(dsName, topic = "CACHE")
}
\description{
读取数据集
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
