% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_dir}
\alias{remove_dir}
\title{删除文件夹}
\usage{
remove_dir(path)
}
\description{
删除文件夹
}
\seealso{
Other utils function: 
\code{\link{\%empty\%}()},
\code{\link{\%error\%}()},
\code{\link{\%nin\%}()},
\code{\link{\%not-empty\%}()},
\code{\link{\%not-regex\%}()},
\code{\link{\%regex\%}()},
\code{\link{create_dir}()},
\code{\link{gen_batchNum}()}
}
\concept{utils function}
