% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_build}
\alias{risk_data_build}
\title{生成疑点数据}
\usage{
risk_data_build(
  modelId,
  riskDataName = "__RISK_DATA__",
  cacheTopic = "CACHE",
  taskTopic = "TASK_DEFINE",
  scriptsTopic = "TASK_SCRIPTS"
)
}
\description{
将筛查后的疑点数据写入数据集
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_read}()},
\code{\link{risk_model_create}()}
}
\concept{risk function}
