% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_clear}
\alias{risk_data_clear}
\title{清理模型遗留数据}
\usage{
risk_data_clear(
  modelName,
  modelGroup,
  dsName = "疑点数据",
  targetTopic = "CACHE"
)
}
\arguments{
\item{modelName}{模型名称}

\item{modelGroup}{模型组名称}

\item{dsName}{疑点数据集名称}

\item{targetTopic}{疑点数据集主题域}
}
\description{
清理模型遗留数据
}
