% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_read}
\alias{risk_data_read}
\title{查看疑点数据}
\usage{
risk_data_read(dataset = "疑点数据", targetTopic = "CACHE")
}
\description{
查看疑点数据
}
\seealso{
Other risk function: 
\code{\link{risk_model_create}()},
\code{\link{risk_model_read}()},
\code{\link{risk_model_run}()}
}
\concept{risk function}
