% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_set}
\alias{risk_data_set}
\title{设置疑点标记}
\usage{
risk_data_set(
  dataId,
  modelGroup,
  flag_new = "__NEW__",
  dsName = "疑点数据",
  targetTopic = "CACHE"
)
}
\arguments{
\item{dataId}{数据ID}

\item{modelGroup}{模型组名称}

\item{dsName}{疑点数据集名称}

\item{targetTopic}{疑点数据集主题域}
}
\description{
设置疑点标记
}
