% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_set_todo}
\alias{risk_data_set_todo}
\title{为疑点数据设置todo标记}
\usage{
risk_data_set_todo(
  d,
  b_todoFlag = TRUE,
  riskDataName = "__RISK_DATA__",
  cacheTopic = "CACHE"
)
}
\description{
为疑点数据设置todo标记
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_read}()},
\code{\link{risk_data_set_done}()},
\code{\link{risk_data_write}()},
\code{\link{risk_model_create}()},
\code{\link{risk_model_search}()}
}
\concept{risk function}
