% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_write}
\alias{risk_data_write}
\title{根据风险模型生成并写入疑点数据}
\usage{
risk_data_write(d, modelId)
}
\arguments{
\item{d}{经过疑点筛查的数据集}

\item{modelId}{风险模型Id，一般被提前配置在任务执行环境中，并做映射}
}
\description{
如果疑点数据已经生成，但未处理，则先清理这些数据
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_read}()},
\code{\link{risk_data_set_done}()},
\code{\link{risk_data_set_todo}()},
\code{\link{task_risk_model_create}()},
\code{\link{task_risk_model_search}()}
}
\concept{risk function}
