% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_filter_add}
\alias{risk_filter_add}
\title{为风险模型增加阈值条件}
\usage{
risk_filter_add(modelId, column, op, value)
}
\description{
允许为数据集增加多个阈值查询条件，缩小筛查范围。
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_generater_add}()},
\code{\link{risk_model_create}()}
}
\concept{risk function}
