% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_generater_add}
\alias{risk_generater_add}
\title{生成疑点数据}
\usage{
risk_generater_add(modelId)
}
\description{
将筛查后的疑点数据写入数据集
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_filter_add}()},
\code{\link{risk_model_create}()}
}
\concept{risk function}
