% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_create}
\alias{risk_model_create}
\title{初始化风险模型}
\usage{
risk_model_create(
  modelName,
  dataset,
  filter = list(list(column = "col_name1", op = ">", value = 0:1)),
  modelGroup = modelName,
  desc = "-",
  author = "-",
  online = FALSE,
  overwrite = FALSE,
  topic = "RISKMODEL"
)
}
\description{
初始化风险模型
}
\seealso{
Other risk function: 
\code{\link{risk_data_read}()},
\code{\link{risk_model_read}()},
\code{\link{risk_model_run}()}
}
\concept{risk function}
