% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_read}
\alias{risk_model_read}
\title{读取风险模型}
\usage{
risk_model_read(topic = "RISKMODEL")
}
\arguments{
\item{topic}{主题名称}
}
\description{
读取风险模型
}
\seealso{
Other risk function: 
\code{\link{risk_data_read}()},
\code{\link{risk_model_create}()},
\code{\link{risk_model_run}()}
}
\concept{risk function}
