% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_run_all}
\alias{risk_model_run_all}
\title{运行所有模型}
\usage{
risk_model_run_all(
  dsName = "疑点数据",
  targetTopic = "CACHE",
  topic = "RISKMODEL"
)
}
\arguments{
\item{dsName}{疑点数据集名称}

\item{targetTopic}{疑点数据集主题域}

\item{topic}{主题名称}
}
\description{
运行所有模型
}
\seealso{
Other risk function: 
\code{\link{risk_data_read}()},
\code{\link{risk_model_all}()},
\code{\link{risk_model_create}()},
\code{\link{risk_model_read}()},
\code{\link{risk_model_run}()}
}
\concept{risk function}
