% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_search}
\alias{risk_model_search}
\title{查找风险模型}
\usage{
risk_model_search(modelMatch = ".*", taskTopic = "TASK_DEFINE")
}
\description{
查找风险模型
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_read}()},
\code{\link{risk_data_write}()},
\code{\link{risk_model_create}()}
}
\concept{risk function}
