% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-folder.R
\name{run_task_scripts}
\alias{run_task_scripts}
\title{执行目标路径下的任务脚本}
\usage{
run_task_scripts(taskScript = "TASK/IMPORT", batch = F)
}
\description{
应当按照脚本顺序执行
}
\seealso{
Other TaskFolder functions: 
\code{\link{ds_import}()},
\code{\link{get_path}()},
\code{\link{get_task_scripts}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{set_topic}()}
}
\concept{TaskFolder functions}
