% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_expr_add}
\alias{script_expr_add}
\title{为任务增加表达式任务脚本}
\usage{
script_expr_add(...)
}
\description{
为任务增加表达式任务脚本
}
\seealso{
Other task-script function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_remove}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()}
}
\concept{task-script function}
