% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_item_add}
\alias{script_item_add}
\title{增加子任务}
\usage{
script_item_add(
  taskName,
  script = NA,
  params = NA,
  globalVars = NA,
  inputAssign = NA,
  outputAssign = NA,
  touchFiles = TRUE,
  taskTopic = "TASK_DEFINE",
  type = "var"
)
}
\arguments{
\item{taskName}{任务标识}

\item{script}{执行脚本或函数名、文件名、目录名}

\item{params}{执行函数参数映射}

\item{globalVars}{设置全局变量}

\item{inputAssign}{使用全局变量或任务参数映射脚本入参}

\item{outputAssign}{将输出结果保存为全局变量}

\item{touchFiles}{类型为file,dir时自动创建脚本文件}

\item{taskTopic}{任务主题存储位置}

\item{type}{可以是string,expr,function,file,dir,var等}
}
\description{
创建任务过程中变量赋值顺序是：globalVars, inputAssign, outputAssign

如果添加的func类型，还应当注意：
\itemize{
\item 要在func脚本中使用globalVars，需要设置inputAssign映射
\item func脚本中的inputAssign设置可能有3种来源：
（1）是\code{@}开头的全局变量；
（2）是\code{globalVars}或\code{outputAssign}输出的全局变量；
（3）是执行\code{task_run}时额外提供的参数。
}
}
\seealso{
Other task-script function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_remove}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()}
}
\concept{task-script function}
