% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_item_remove}
\alias{script_item_remove}
\title{删除任务脚本}
\usage{
script_item_remove(taskId, rowNum, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskId}{任务标识}

\item{rowNum}{要更新的脚本序号}

\item{taskTopic}{任务主题存储位置}
}
\description{
删除任务脚本
}
\seealso{
Other task-define function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()},
\code{\link{task_create}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
