% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_item_update}
\alias{script_item_update}
\title{更新任务脚本}
\usage{
script_item_update(
  taskId,
  rowNum,
  script = NA,
  params = NA,
  globalVars = NA,
  inputAsign = NA,
  outputAsign = NA,
  touchFiles = TRUE,
  taskTopic = "TASK_DEFINE",
  type = "global"
)
}
\arguments{
\item{taskId}{任务标识}

\item{rowNum}{要更新的脚本序号}

\item{script}{执行脚本或函数名、文件名、目录名}

\item{params}{执行函数参数映射}

\item{globalVars}{设置全局变量}

\item{inputAsign}{针对function和gali类型，使用执行环境内变量映射入参}

\item{outputAsign}{保存子任务输出}

\item{touchFiles}{类型为file,dir时自动创建脚本文件}

\item{taskTopic}{任务主题存储位置}

\item{type}{可以是string,expr,function,gali,file,dir,global等}
}
\description{
更新任务脚本
}
\seealso{
Other task-script function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_remove}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()}
}
\concept{task-script function}
