% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_var_update}
\alias{script_var_update}
\title{更新为设置全局变量任务脚本}
\usage{
script_var_update(...)
}
\description{
更新为设置全局变量任务脚本
}
\seealso{
Other task-script function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_remove}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()}
}
\concept{task-script function}
