% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_topic}
\alias{set_topic}
\title{设置主题目录}
\usage{
set_topic(topic, path)
}
\arguments{
\item{topic}{任务主题}

\item{path}{文件夹位置}
}
\description{
设置主题目录
}
\seealso{
Other config functions: 
\code{\link{get_config}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{load_config}()}
}
\concept{config functions}
