% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_input_assign.R
\name{sm_input_assign_ui}
\alias{sm_input_assign_ui}
\title{Shiny Modules UI - 任务脚本入参映射}
\usage{
sm_input_assign_ui(id, inputAssign)
}
\description{
Shiny Modules UI - 任务脚本入参映射
}
\seealso{
Other Shiny Modules functions: 
\code{\link{sm_global_vars_server}()},
\code{\link{sm_global_vars_ui}()},
\code{\link{sm_input_assign_server}()},
\code{\link{sm_output_assign_server}()},
\code{\link{sm_output_assign_ui}()},
\code{\link{sm_params_server}()},
\code{\link{sm_params_ui}()},
\code{\link{sm_preview_server}()},
\code{\link{sm_preview_ui}()},
\code{\link{sm_scripts_server}()},
\code{\link{sm_scripts_ui}()},
\code{\link{sm_task_server}()},
\code{\link{sm_task_ui}()}
}
\concept{Shiny Modules functions}
