% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_params.R
\name{sm_params_ui}
\alias{sm_params_ui}
\title{Shiny Modules UI - 任务参数}
\usage{
sm_params_ui(id, params)
}
\description{
Shiny Modules UI - 任务参数
}
\seealso{
Other Shiny Modules functions: 
\code{\link{sm_params_server}()},
\code{\link{sm_preview_server}()},
\code{\link{sm_preview_ui}()},
\code{\link{sm_scripts_server}()},
\code{\link{sm_scripts_ui}()},
\code{\link{sm_task_server}()},
\code{\link{sm_task_ui}()}
}
\concept{Shiny Modules functions}
