% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_preview.R
\name{sm_preview_server}
\alias{sm_preview_server}
\title{Shiny Modules Server - 结果预览}
\usage{
sm_preview_server(id, content, title = "")
}
\description{
Shiny Modules Server - 结果预览
}
\seealso{
Other Shiny Modules functions: 
\code{\link{sm_params_server}()},
\code{\link{sm_params_ui}()},
\code{\link{sm_preview_ui}()},
\code{\link{sm_scripts_server}()},
\code{\link{sm_scripts_ui}()},
\code{\link{sm_task_server}()},
\code{\link{sm_task_ui}()}
}
\concept{Shiny Modules functions}
