% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_scripts.R
\name{sm_scripts_ui}
\alias{sm_scripts_ui}
\title{Shiny Modules UI - 任务脚本}
\usage{
sm_scripts_ui(id, scriptItems)
}
\description{
Shiny Modules UI - 任务脚本
}
\seealso{
Other Shiny Modules functions: 
\code{\link{sm_params_server}()},
\code{\link{sm_params_ui}()},
\code{\link{sm_preview_server}()},
\code{\link{sm_preview_ui}()},
\code{\link{sm_scripts_server}()},
\code{\link{sm_task_server}()},
\code{\link{sm_task_ui}()}
}
\concept{Shiny Modules functions}
