% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_task.R
\name{sm_task_ui}
\alias{sm_task_ui}
\title{Shiny Modules UI - 任务管理}
\usage{
sm_task_ui(id)
}
\description{
Shiny Modules UI - 任务管理
}
\seealso{
Other Shiny Modules functions: 
\code{\link{sm_params_server}()},
\code{\link{sm_params_ui}()},
\code{\link{sm_preview_server}()},
\code{\link{sm_preview_ui}()},
\code{\link{sm_scripts_server}()},
\code{\link{sm_scripts_ui}()},
\code{\link{sm_task_server}()}
}
\concept{Shiny Modules functions}
