% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_all}
\alias{state_all}
\title{列举所有状态数据集}
\usage{
state_all(...)
}
\description{
与\code{ds_all}相同，但主题域默认为STATE
}
\seealso{
Other state function: 
\code{\link{state_read}()},
\code{\link{state_write}()}
}
\concept{state function}
