% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_read}
\alias{state_read}
\title{读取状态信息}
\usage{
state_read(stateName, topic = "STATE")
}
\arguments{
\item{stateName}{状态数据集名称}
}
\description{
读取状态信息
}
\seealso{
Other state function: 
\code{\link{state_all}()},
\code{\link{state_write}()}
}
\concept{state function}
