% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-define.R
\name{task_add}
\alias{task_add}
\title{增加子任务}
\usage{
task_add(
  taskId,
  taskScript,
  params = list(NULL),
  scriptType = "string",
  taskTopic = "TASK_DEFINE"
)
}
\arguments{
\item{taskId}{任务标识}

\item{taskScript}{子任务的执行路径}

\item{params}{子任务的参数设置}

\item{scriptType}{可以是string,file或folder}

\item{taskTopic}{保存任务定义的存储主题文件夹}

\item{family}{task-define function}
}
\description{
增加子任务
}
