% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_cancel_snap}
\alias{task_cancel_snap}
\title{取消编辑模式}
\usage{
task_cancel_snap(...)
}
\description{
取消编辑模式
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
