% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-DT.R
\name{task_DT}
\alias{task_DT}
\title{数据表}
\usage{
task_DT(
  d,
  rownames = FALSE,
  selection = "none",
  extensions = "Responsive",
  class = "white-space: nowrap",
  options = list(responsive = TRUE),
  ...
)
}
\description{
数据表
}
\examples{
mtcars |> task_DT()
}
\seealso{
Other task-lib functions: 
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
