% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_exists}
\alias{task_exists}
\title{判断任务定义文件是否存在}
\usage{
task_exists(taskName, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskName}{任务标识}

\item{taskTopic}{主题域}
}
\description{
判断任务定义文件是否存在
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
