% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_file_add}
\alias{task_file_add}
\title{为任务增加文件脚本子任务}
\usage{
task_file_add(...)
}
\description{
为任务增加文件脚本子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_empty_add}()},
\code{\link{task_expr_add}()},
\code{\link{task_func_add}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()}
}
\concept{task-define function}
