% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_files}
\alias{task_files}
\title{查看执行计划}
\usage{
task_files(taskTopic = "TASK/BUILD", taskFolder = "", glob = "*.R")
}
\arguments{
\item{taskTopic}{脚本文件夹主题}

\item{taskFolder}{执行脚本文件的目录}

\item{glob}{要执行的源文件默认以.R结尾}
}
\description{
按执行顺序罗列需要执行的脚本。
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_run}()}
}
\concept{task functions}
