% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_func_add}
\alias{task_func_add}
\title{为任务增加函数子任务}
\usage{
task_func_add(...)
}
\description{
为任务增加函数子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_expr_add}()},
\code{\link{task_file_add}()},
\code{\link{task_global_add}()},
\code{\link{task_item_add}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()}
}
\concept{task-define function}
