% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_gali_add}
\alias{task_gali_add}
\title{为任务增加gali函数子任务}
\usage{
task_gali_add(...)
}
\description{
为任务增加gali函数子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run_gali}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
