% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_global_add}
\alias{task_global_add}
\title{设置全局变量}
\usage{
task_global_add(...)
}
\description{
设置全局变量
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_expr_add}()},
\code{\link{task_file_add}()},
\code{\link{task_func_add}()},
\code{\link{task_item_add}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()}
}
\concept{task-define function}
