% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_id}
\alias{task_id}
\title{从任务名称转换任务ID}
\usage{
task_id(taskName, taskTopic = "TASK_DEFINE")
}
\description{
从任务名称转换任务ID
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_exists}()},
\code{\link{task_name}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
