% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-import.R
\name{task_import_define}
\alias{task_import_define}
\title{创建导入任务的目录结构和脚本文件}
\usage{
task_import_define(taskFolder, taskTopic = "TASK_SCRIPTS")
}
\description{
约定taskFolder与导入素材的目录结构对应。

当taskFolder与导入素材匹配后，就生成一个任务，
这将会执行taskFolder内所有R脚本文件（包括子文件夹）。
}
\seealso{
Other task-import function: 
\code{\link{task_import_init}()},
\code{\link{task_import_run}()},
\code{\link{task_import_scan}()}
}
\concept{task-import function}
