% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-import.R
\name{task_import_init}
\alias{task_import_init}
\title{初始化队列数据集}
\usage{
task_import_init(dsName = "__IMPORT_TASK__", cacheTopic = "CACHE")
}
\description{
处理任务的队列
}
\seealso{
Other task-import function: 
\code{\link{task_import_define}()},
\code{\link{task_import_run}()},
\code{\link{task_import_scan}()}
}
\concept{task-import function}
