% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-import.R
\name{task_import_run}
\alias{task_import_run}
\title{处理导入任务}
\usage{
task_import_run()
}
\description{
处理导入任务
}
\seealso{
Other task-import function: 
\code{\link{task_import_define}()},
\code{\link{task_import_init}()},
\code{\link{task_import_scan}()}
}
\concept{task-import function}
