% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-import.R
\name{task_import_scan}
\alias{task_import_scan}
\title{扫描需要导入的新任务}
\usage{
task_import_scan(
  dsName = "__IMPORT_TASK__",
  cacheTopic = "CACHE",
  importTopic = "IMPORT",
  taskTopic = "TASK_SCRIPTS"
)
}
\description{
扫描需要导入的新任务
}
\seealso{
Other task-import function: 
\code{\link{task_import_define}()},
\code{\link{task_import_init}()},
\code{\link{task_import_run}()}
}
\concept{task-import function}
