% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_item_exchange}
\alias{task_item_exchange}
\title{调整任务脚本顺序}
\usage{
task_item_exchange(taskId, rowNumA, rowNumB, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskId}{任务标识}

\item{taskTopic}{任务主题存储位置}

\item{rowNumOld}{脚本旧序号}

\item{rowNumNew}{脚本新序号}
}
\description{
调整任务脚本顺序
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_dir_update}()},
\code{\link{task_expr_add}()},
\code{\link{task_expr_update}()},
\code{\link{task_file_add}()},
\code{\link{task_file_update}()},
\code{\link{task_func_add}()},
\code{\link{task_func_update}()},
\code{\link{task_global_add}()},
\code{\link{task_global_update}()},
\code{\link{task_item_add}()},
\code{\link{task_item_remove}()},
\code{\link{task_item_update}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()},
\code{\link{task_string_update}()},
\code{\link{task_update}()}
}
\concept{task-define function}
