% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_name}
\alias{task_name}
\title{从任务ID转换任务名称}
\usage{
task_name(taskId, taskTopic = "TASK_DEFINE")
}
\description{
该操作可能在内存中做大量重复计算，因此做缓存处理
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_exists}()},
\code{\link{task_id}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
