% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_offline}
\alias{task_offline}
\title{任务下线}
\usage{
task_offline(...)
}
\description{
任务下线
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
