% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_online}
\alias{task_online}
\title{任务上线}
\usage{
task_online(...)
}
\description{
任务上线
}
\seealso{
Other task-define function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_remove}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()},
\code{\link{task_create}()},
\code{\link{task_offline}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
