% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_params_assign}
\alias{task_params_assign}
\title{提取任务执行的入参映射}
\usage{
task_params_assign(taskName, taskTopic = "TASK_DEFINE")
}
\description{
\code{\link{task_run}}函数执行时允许携带额外参数，
这可以从func类型脚本中的\code{inputAssign}自动提取。

提取的\code{inputAssign}设置时，自动过滤映射值为@开头的内置参数，
或映射值为outputAssign或globalVars中已经输出的参数。

返回值中会包含入参映射所在函数的元数据描述。

如果脚本中设置过该参数，则将其设为默认值。
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_exists}()},
\code{\link{task_id}()},
\code{\link{task_name}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
