% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_area}
\alias{task_plotly_area}
\title{面积图}
\usage{
task_plotly_area(d, x, y, fill = "tozeroy", color = "red", shape = "spline")
}
\description{
面积图
}
\examples{
mtcars |>
  count(cyl) |>
  task_plotly_area(x = "cyl", y = "n")
  
mtcars |> group_by(cyl) |>
  summarise(displ = mean(disp)) |>
  task_plotly_area(x = "cyl", y = "displ")
  
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
