% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_bar}
\alias{task_plotly_bar}
\title{柱图}
\usage{
task_plotly_bar(d, x, y)
}
\description{
柱图
}
\examples{
mtcars |>
  count(cyl) |>
  task_plotly_bar(x = "cyl", y = "n")

mpg |> count(class) |>
  mutate(class = forcats::fct_reorder(class, n, .desc = TRUE)) |>
  task_plotly_bar(x = "class", y = "n")
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
