% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_barpolar}
\alias{task_plotly_barpolar}
\title{极坐标柱}
\usage{
task_plotly_barpolar(
  d,
  theta,
  r,
  color = NULL,
  bargap = 0,
  hole = 0.05,
  direction = "clockwise"
)
}
\description{
极坐标柱
}
\examples{
mtcars |>
  count(cyl) |>
  mutate(cyl = sprintf("CYL: \%d", cyl)) |>
  task_plotly_barpolar(theta = "cyl", r = "n")
  
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
