% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_hist}
\alias{task_plotly_hist}
\title{直方图}
\usage{
task_plotly_hist(d, x, color = "brown", xtitle = NULL, ytitle = NULL)
}
\description{
直方图
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
