% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_line}
\alias{task_plotly_line}
\title{线图}
\usage{
task_plotly_line(
  d,
  x,
  y,
  fillcolor = "red",
  linecolor = "rgb(205, 12, 24)",
  shape = "spline"
)
}
\description{
线图
}
\examples{

}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
