% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_marker}
\alias{task_plotly_marker}
\title{散点图}
\usage{
task_plotly_marker(d, x, y, alpha = 0.2, name = NULL)
}
\description{
散点图
}
\examples{
mpg |> task_plotly_marker("cty", "hwy")
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_pie2}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
