% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_pie}
\alias{task_plotly_pie}
\title{饼图}
\usage{
task_plotly_pie(d, value, label = NULL, hole = 0.4)
}
\description{
饼图
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie2}()}
}
\concept{task-lib functions}
