% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{task_plotly_pie2}
\alias{task_plotly_pie2}
\title{饼图}
\usage{
task_plotly_pie2(d, value, label = NULL, pull = 0, hole = 0.25)
}
\description{
饼图
}
\examples{
mtcars |> count(cyl) |>
 task_plotly_pie2(value = "n", pull = c(0.1, 0, 0))
}
\seealso{
Other task-lib functions: 
\code{\link{task_DT}()},
\code{\link{task_plotly_area}()},
\code{\link{task_plotly_barpolar}()},
\code{\link{task_plotly_bar}()},
\code{\link{task_plotly_hist}()},
\code{\link{task_plotly_line}()},
\code{\link{task_plotly_marker}()},
\code{\link{task_plotly_pie}()}
}
\concept{task-lib functions}
