% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_init}
\alias{task_queue_init}
\title{初始化队列数据集}
\usage{
task_queue_init(dsName = "__TASK_QUEUE__", cacheTopic = "CACHE")
}
\description{
批处理任务的队列
}
\seealso{
Other queue function: 
\code{\link{task_queue_item}()},
\code{\link{task_queue_param_from_yaml}()},
\code{\link{task_queue_param_to_yaml}()},
\code{\link{task_queue_search}()}
}
\concept{queue function}
