% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_item}
\alias{task_queue_item}
\title{构造队列中的一条数据}
\usage{
task_queue_item(
  taskId,
  params,
  taskType = "__TYPE_UNKNOWN__",
  taskTopic = "CACHE",
  id = gen_batchNum(),
  runLevel = 500L
)
}
\description{
构造队列中的一条数据
}
\seealso{
Other queue function: 
\code{\link{task_queue_all}()},
\code{\link{task_queue_init}()},
\code{\link{task_queue_param_from_yaml}()},
\code{\link{task_queue_param_to_yaml}()},
\code{\link{task_queue_run}()},
\code{\link{task_queue_todo}()}
}
\concept{queue function}
