% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_param_from_yaml}
\alias{task_queue_param_from_yaml}
\title{从压马路格式提取队列参数}
\usage{
task_queue_param_from_yaml(ymlParams)
}
\description{
从压马路格式提取队列参数
}
\seealso{
Other queue function: 
\code{\link{task_queue_all}()},
\code{\link{task_queue_init}()},
\code{\link{task_queue_item}()},
\code{\link{task_queue_param_to_yaml}()},
\code{\link{task_queue_run}()},
\code{\link{task_queue_todo}()}
}
\concept{queue function}
