% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_param_to_yaml}
\alias{task_queue_param_to_yaml}
\title{转换队列参数为yaml格式}
\usage{
task_queue_param_to_yaml(params)
}
\description{
转换队列参数为yaml格式
}
\seealso{
Other queue function: 
\code{\link{task_queue_init}()},
\code{\link{task_queue_item}()},
\code{\link{task_queue_param_from_yaml}()},
\code{\link{task_queue_search}()}
}
\concept{queue function}
