% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_run}
\alias{task_queue_run}
\title{批量执行队列任务}
\usage{
task_queue_run(
  dsQueue,
  dsName = "__TASK_QUEUE__",
  cacheTopic = "CACHE",
  runMode = "in-process"
)
}
\description{
按批次执行队列内的任务。

相同批次的任务按优先级大小依次执行；
执行结果统一更新到队列。

如果批任务中部分任务失败，则整体批次任务暂停，
批次中的所有任务信息都不更新。
}
\seealso{
Other queue function: 
\code{\link{task_queue_all}()},
\code{\link{task_queue_init}()},
\code{\link{task_queue_item}()},
\code{\link{task_queue_param_from_yaml}()},
\code{\link{task_queue_param_to_yaml}()},
\code{\link{task_queue_todo}()}
}
\concept{queue function}
