% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_todo}
\alias{task_queue_todo}
\title{待执行的队列任务}
\usage{
task_queue_todo(
  taskTypes = c("__IMPORT__", "__BUILD__", "__RISK__", "__SUMMARISE__"),
  dsName = "__TASK_QUEUE__",
  cacheTopic = "CACHE"
)
}
\description{
待执行的队列任务
}
\seealso{
Other queue function: 
\code{\link{task_queue_init}()},
\code{\link{task_queue_item}()},
\code{\link{task_queue_param_from_yaml}()},
\code{\link{task_queue_param_to_yaml}()},
\code{\link{task_queue_run}()}
}
\concept{queue function}
