% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_read}
\alias{task_read}
\title{查看执行计划}
\usage{
task_read(taskScript, glob = "*.R")
}
\arguments{
\item{taskScript}{脚本文件夹主题}

\item{glob}{要执行的源文件默认以.R结尾}
}
\description{
按执行顺序罗列需要执行的脚本
}
\seealso{
Other TaskFolder functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{set_topic}()},
\code{\link{task_run}()}
}
\concept{TaskFolder functions}
