% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run0}
\alias{task_run0}
\title{按内容运行任务}
\usage{
task_run0(taskItems, runMode = "in-process", ...)
}
\arguments{
\item{runMode}{运行模式（默认为进程内执行，改为r或r_bg为子进程执行）}

\item{task}{任务}
}
\description{
按内容运行任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
