% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run_dir}
\alias{task_run_dir}
\title{运行任务文件夹}
\usage{
task_run_dir(
  taskDir,
  params = list(NULL),
  scriptsTopic = "TASK_SCRIPTS",
  runMode = "in-process",
  ...
)
}
\arguments{
\item{taskDir}{任务文件夹路径}
}
\description{
运行任务文件夹
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_script_dir_create}()},
\code{\link{task_script_file_create}()},
\code{\link{task_search}()}
}
\concept{task-define function}
