% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run_file}
\alias{task_run_file}
\title{快速运行脚本文件}
\usage{
task_run_file(
  taskFile,
  params = list(NULL),
  scriptsTopic = "TASK_SCRIPTS",
  runMode = "in-process",
  ...
)
}
\arguments{
\item{taskFile}{任务文件路径}
}
\description{
快速运行脚本文件
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
