% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run_string}
\alias{task_run_string}
\title{快速运行脚本字符串}
\usage{
task_run_string(taskString, params = list(NULL), runMode = "in-process", ...)
}
\arguments{
\item{taskString}{脚本字符串}
}
\description{
快速运行脚本字符串
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
