% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_script_file_create}
\alias{task_script_file_create}
\title{创建脚本文件}
\usage{
task_script_file_create(scriptFile, scriptsTopic = "TASK_SCRIPTS")
}
\description{
创建脚本文件
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_script_dir_create}()},
\code{\link{task_search}()}
}
\concept{task-define function}
