% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_string_add}
\alias{task_string_add}
\title{为任务增加字符串脚本子任务}
\usage{
task_string_add(...)
}
\description{
为任务增加字符串脚本子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_empty_add}()},
\code{\link{task_expr_add}()},
\code{\link{task_file_add}()},
\code{\link{task_func_add}()},
\code{\link{task_gali_add}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
