% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_update}
\alias{task_update}
\title{更新任务配置}
\usage{
task_update(taskId, ..., force = FALSE, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskId}{任务唯一标识，每个\code{taskId}会保存为一个独立文件}

\item{taskTopic}{保存任务定义的存储主题文件夹}

\item{taskType}{任务类型}

\item{online}{如果任务下线，在推荐可执行任务时将忽略}

\item{desc}{任务描述}
}
\description{
更新任务配置
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_dir_update}()},
\code{\link{task_expr_add}()},
\code{\link{task_expr_update}()},
\code{\link{task_file_add}()},
\code{\link{task_file_update}()},
\code{\link{task_func_add}()},
\code{\link{task_func_update}()},
\code{\link{task_global_add}()},
\code{\link{task_global_update}()},
\code{\link{task_item_add}()},
\code{\link{task_item_exchange}()},
\code{\link{task_item_remove}()},
\code{\link{task_item_update}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()},
\code{\link{task_string_update}()}
}
\concept{task-define function}
