% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china-workdays.R
\name{workdays_patch}
\alias{workdays_patch}
\title{计算工作日}
\usage{
workdays_patch(
  d,
  fromDayName,
  toDayName,
  fromDay = "2019-01-01",
  toDay = "2022-12-31"
)
}
\arguments{
\item{d}{要补充字段的数据集}

\item{fromDayName}{开始日期字段名}

\item{toDayName}{截止日期字段名}

\item{fromDay}{开始日期}

\item{toDay}{截止日期}
}
\description{
计算数据集中指定两列占用的工作日数量
}
