% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{write_dataset}
\alias{write_dataset}
\title{写入ApachheParquet文件集}
\usage{
write_dataset(d, dsName, topic = "CACHE", partColumns = c(), keyColumns = c())
}
\description{
更新受影响的分区
}
\details{
默认为更新模式；如果仅直接覆写数据，应先删除数据集

如果不设置关键列，则为追加模式；否则按关键列替换
}
\seealso{
Other dataset function: 
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{last_affected_parts}()},
\code{\link{read_affected_parts}()},
\code{\link{read_dataset}()}
}
\concept{dataset function}
