% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-state.R
\name{write_state}
\alias{write_state}
\title{写入状态信息}
\usage{
write_state(stateName, dataNew)
}
\arguments{
\item{taskName}{任务名称，一般为“导入“，”构建“，”上传“等}

\item{title}{子主题名称，可以是导入文件夹或某日构建等，也可以区分取数任务和取数明细}

\item{datasetName}{数据集名称}

\item{rowID}{数据行ID，一般为唯一关键列，可用于比对数据是否处理过等}

\item{flag}{状态标志，方便读取状态信息时过滤}

\item{info}{状态信息的详细描述}
}
\description{
状态信息一定保存在STATE主题的目录下：{STATE}/{taskName}
}
\details{
应注意状态记录并非覆盖写入，而是追加，且一般不做删除或修改
有明确约定的状态数据包括：

1、标记处理完成的任务文件夹时，stateName = "\strong{TASK_FOLDER}"
2、写入parquet数据集文件时，stateName = "\strong{WRITE_DATASET}"
}
