% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_init}
\alias{config_init}
\title{初始化配置项}
\usage{
config_init(path, yml = "config.yml", option = list())
}
\arguments{
\item{path}{YAML配置文件目录}

\item{yml}{默认为config.yml}

\item{option}{配置项}
}
\description{
初始化配置项
}
\seealso{
Other config functions: 
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{config_yaml}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()}
}
\concept{config functions}
