% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_filter}
\alias{dp_filter}
\title{定义数据过滤任务}
\usage{
dp_filter(s_column, op_name, fn_value = list(NULL), s_dataName = "@result")
}
\arguments{
\item{taskId}{任务ID}

\item{column}{列名}

\item{op}{阈值范围判断符号}

\item{value}{阈值}

\item{dataName}{内存中的数据框名称，默认为 @result}

\item{taskTopic}{任务定义的主题文件夹}
}
\description{
这个任务的目标是通过配置项实现函数\code{\link[dplyr:filter]{dplyr::filter()}}的大部分功能。

允许为数据集增加多个阈值查询条件，缩小筛查范围。

taskId、column、op、value等参数构造唯一的dp_filter表达式，
这将允许从UI生成或还原该操作。

只要包括以下逻辑判断：
\itemize{
\item >, <, >=, <=, ==, !=
\item \%in\%, \%nin\%
\item \%regex\%, \%not-regex\%
\item #time# >, #time# <
\item #date# >, #date# <
}
}
\seealso{
Other data-plyr function: 
\code{\link{dp_arrange}()},
\code{\link{dp_collect}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_rename}()},
\code{\link{dp_select}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
