% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plyr.R
\name{dp_select}
\alias{dp_select}
\title{选择列，支持惰性计算}
\usage{
dp_select(
  sn_columns = list(),
  b_everything = FALSE,
  s_regex = NULL,
  s_dataName = "@result"
)
}
\description{
选择列，支持惰性计算
}
\seealso{
Other data-plyr function: 
\code{\link{dp_arrange}()},
\code{\link{dp_collect}()},
\code{\link{dp_filter}()},
\code{\link{dp_head}()},
\code{\link{dp_n_max}()},
\code{\link{dp_n_min}()},
\code{\link{dp_read}()},
\code{\link{dp_rename}()},
\code{\link{dp_tail}()}
}
\concept{data-plyr function}
