% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_all}
\alias{ds_all}
\title{列举所有数据集}
\usage{
ds_all(topic = "CACHE")
}
\arguments{
\item{topic}{主题域}
}
\description{
列举所有数据集
}
\seealso{
Other dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_append}()},
\code{\link{ds_arrange}()},
\code{\link{ds_as_datetime}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_delete}()},
\code{\link{ds_demo}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_distinct}()},
\code{\link{ds_drop}()},
\code{\link{ds_dup_count}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_init}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_rename}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
