% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_as_datetime}
\alias{ds_as_datetime}
\title{转换字符串为时间日期类型}
\usage{
ds_as_datetime(ds, timestampColumn, datetimeColumn, tzone = "Asia/Shanghai")
}
\description{
转换字符串为时间日期类型
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
