% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-parquet.R
\name{ds_as_unique}
\alias{ds_as_unique}
\title{去除重复行}
\usage{
ds_as_unique(ds, keyColumns)
}
\arguments{
\item{ds}{要确认的数据集}

\item{columns}{要确认的列名或其向量、列表}
}
\description{
按键值列去除重复行。
}
\seealso{
Other dataset function: 
\code{\link{all_dataset}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{last_affected_parts}()},
\code{\link{read_affected_parts}()},
\code{\link{read_dataset}()},
\code{\link{write_dataset}()}
}
\concept{dataset function}
