% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_delete}
\alias{ds_delete}
\title{批量追加删除的数据}
\usage{
ds_delete(d, dsName, topic = "CACHE")
}
\arguments{
\item{d}{要删除的数据}

\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}
}
\description{
以追加标记文件的形式删除数据（不做物理删除）
}
\seealso{
Other dataset function: 
\code{\link{ds_add_count}()},
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_arrange}()},
\code{\link{ds_as_datetime}()},
\code{\link{ds_as_deleted}()},
\code{\link{ds_as_from}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_collect}()},
\code{\link{ds_count}()},
\code{\link{ds_demo}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_drop}()},
\code{\link{ds_exists}()},
\code{\link{ds_files}()},
\code{\link{ds_filter}()},
\code{\link{ds_head}()},
\code{\link{ds_init}()},
\code{\link{ds_n_max}()},
\code{\link{ds_n_min}()},
\code{\link{ds_read0}()},
\code{\link{ds_read}()},
\code{\link{ds_rename}()},
\code{\link{ds_schema_obj}()},
\code{\link{ds_schema}()},
\code{\link{ds_select}()},
\code{\link{ds_tail}()},
\code{\link{ds_type}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
