% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_init}
\alias{ds_init}
\title{初始化数据集配置}
\usage{
ds_init(
  dsName,
  topic = "CACHE",
  partColumns = c(),
  keyColumns = c(),
  suggestedColumns = c(),
  titleColumn = c(),
  desc = "-"
)
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}

\item{partColumns}{支持分区存储，支持多列}

\item{keyColumns}{要求在所有行中具有唯一值的列向量，支持多列}

\item{suggestedColumns}{查看数据时推荐的显示列向量，支持多列，}

\item{titleColumn}{标题列，不支持多列}

\item{desc}{对数据集的额外描述}
}
\description{
所有数据集使用@action作为第一层分区：
\strong{APPEND}, \strong{DELETE}, \strong{ARCHIVE}
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_last_affected}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
