% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_schema}
\alias{ds_schema}
\title{查看数据集架构}
\usage{
ds_schema(ds)
}
\arguments{
\item{ds}{数据框}
}
\description{
查看数据集架构时，必须包含数据
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_append}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_delete}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_files}()},
\code{\link{ds_init}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_write}()}
}
\concept{dataset function}
