% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_write}
\alias{ds_write}
\title{写入ApachheParquet文件集}
\usage{
ds_write(
  d,
  dsName,
  topic = "CACHE",
  partColumns = c(),
  keyColumns = c(),
  suggestedColumns = c(),
  titleColumn = c(),
  desc = "-",
  mode = "update"
)
}
\arguments{
\item{d}{要写入的数据}

\item{dsName}{数据集名称}

\item{topic}{数据集保存的主题目录，默认为CACHE}

\item{partColumns}{照此进行分区存储，支持多列}

\item{keyColumns}{要求在所有行中具有唯一值的列向量，支持多列}

\item{suggestedColumns}{查看数据时推荐的显示列向量，支持多列，}

\item{titleColumn}{标题列，不支持多列}

\item{desc}{对数据集的额外描述}

\item{mode}{当主键重复时，update模式（默认）使用新数据，append模式保留旧数据}
}
\description{
更新受影响的分区
}
\details{
默认为更新模式；如果仅直接覆写数据，应先删除数据集

如果不设置关键列，则为追加模式；否则按关键列替换
}
\seealso{
Other dataset function: 
\code{\link{ds_all}()},
\code{\link{ds_as_unique}()},
\code{\link{ds_compare_schema}()},
\code{\link{ds_diff_dataset}()},
\code{\link{ds_diff_schema}()},
\code{\link{ds_last_affected}()},
\code{\link{ds_read_affected}()},
\code{\link{ds_read}()},
\code{\link{ds_remove_path}()},
\code{\link{ds_schema}()}
}
\concept{dataset function}
