% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-yaml.R
\name{ds_yaml}
\alias{ds_yaml}
\title{查看数据集元数据}
\usage{
ds_yaml(dsName, topic = "CACHE")
}
\arguments{
\item{dsName}{数据集名称}

\item{topic}{主题域}
}
\description{
查看数据集元数据
}
\seealso{
Other metadata function: 
\code{\link{ds_write_yaml}()}
}
\concept{metadata function}
