% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-DT.R
\name{gali_DT}
\alias{gali_DT}
\title{数据表}
\usage{
gali_DT(
  `@ds`,
  b_rownames = FALSE,
  e_selection = "none",
  e_extensions = "Responsive",
  e_class = "white-space: nowrap",
  e_options = list(responsive = TRUE),
  ...
)
}
\description{
数据表
}
\examples{
mtcars |> gali_DT()
}
\concept{lib-DT functions}
