% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gali-dataset.R
\name{gali_read}
\alias{gali_read}
\title{读取数据集}
\usage{
gali_read(s_dsName)
}
\description{
读取数据集
}
\seealso{
Other gali-dataset function: 
\code{\link{gali_dataset_arrange}()},
\code{\link{gali_dataset_collect}()},
\code{\link{gali_dataset_filter}()},
\code{\link{gali_dataset_head}()},
\code{\link{gali_dataset_n_max}()},
\code{\link{gali_dataset_n_min}()},
\code{\link{gali_dataset_rename}()},
\code{\link{gali_dataset_select}()},
\code{\link{gali_dataset_tail}()},
\code{\link{gali_write}()}
}
\concept{gali-dataset function}
