% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-folder.R
\name{get_path}
\alias{get_path}
\title{根据配置构造数据路径}
\usage{
get_path(topic, ...)
}
\arguments{
\item{topic}{任务主题}
}
\description{
根据配置构造数据路径
}
\seealso{
Other TaskFolder functions: 
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{set_topic}()}
}
\concept{TaskFolder functions}
