% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_path}
\alias{get_path}
\title{根据配置构造数据路径}
\usage{
get_path(topic, ...)
}
\arguments{
\item{topic}{任务主题}
}
\description{
根据配置构造数据路径
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{config_yaml}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()}
}
\concept{config functions}
