% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_topic}
\alias{get_topic}
\title{获取配置}
\usage{
get_topic(topic)
}
\arguments{
\item{topic}{任务主题}
}
\description{
获取配置
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{config_yaml}()},
\code{\link{get_path}()},
\code{\link{get_topics}()}
}
\concept{config functions}
