% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{get_topic}
\alias{get_topic}
\title{获取配置}
\usage{
get_topic(topic)
}
\arguments{
\item{topic}{任务主题}
}
\description{
获取配置
}
\seealso{
Other TaskFolder functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{set_topic}()},
\code{\link{task_read}()},
\code{\link{task_run}()}
}
\concept{TaskFolder functions}
