% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{get_topic}
\alias{get_topic}
\title{获取配置}
\usage{
get_topic(topic)
}
\arguments{
\item{topic}{任务主题}
}
\description{
获取配置
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{load_config}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
