% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_topics}
\alias{get_topics}
\title{列举所有配置项}
\usage{
get_topics()
}
\description{
列举所有配置项
}
\seealso{
Other config functions: 
\code{\link{config_init}()},
\code{\link{config_load}()},
\code{\link{config_write}()},
\code{\link{config_yaml}()},
\code{\link{get_path}()},
\code{\link{get_topic}()}
}
\concept{config functions}
