% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{get_topics}
\alias{get_topics}
\title{获得所有主题}
\usage{
get_topics()
}
\description{
获得所有主题
}
\seealso{
Other task functions: 
\code{\link{get_path}()},
\code{\link{get_topic}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{set_topic}()},
\code{\link{task_read}()},
\code{\link{task_run}()}
}
\concept{task functions}
