% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_topics}
\alias{get_topics}
\title{获得所有主题}
\usage{
get_topics()
}
\description{
获得所有主题
}
\seealso{
Other config functions: 
\code{\link{get_config}()},
\code{\link{get_path}()},
\code{\link{get_topic}()},
\code{\link{load_config}()},
\code{\link{set_topic}()}
}
\concept{config functions}
