% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{import_redo}
\alias{import_redo}
\title{手工指定要处理的任务文件夹}
\usage{
import_redo(
  todo = c(),
  importTopic = "IMPORT",
  taskTopic = "TASK/IMPORT",
  taskFolder = ""
)
}
\description{
手工指定要处理的任务文件夹
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_todo}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
