% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{加载配置文件}
\usage{
load_config(yml_file)
}
\arguments{
\item{yml_file}{YAML配置文件}
}
\description{
批量执行\code{set_topic}任务
}
\details{
使用\code{ROOT_PATH}时有一个关键约定：
配置项必须以\code{./}开头，才能使用\code{ROOT_PATH}扩展其路径；
否则，将被视为独立配置名。
}
\seealso{
Other config functions: 
\code{\link{get_config}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{set_topic}()}
}
\concept{config functions}
