% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_area}
\alias{plot_area}
\title{面积图}
\usage{
plot_area(d, e_x, e_y, c_fill = "tozeroy", c_color = "red", e_shape = "spline")
}
\description{
面积图
}
\examples{
mtcars |>
  count(cyl) |>
  plot_area(x = "cyl", y = "n")
  
mtcars |> group_by(cyl) |>
  summarise(displ = mean(disp)) |>
  gali_plot_area(e_x = "cyl", e_y = "displ")
  
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_init}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
