% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{plot_area}
\alias{plot_area}
\title{面积图}
\usage{
plot_area(d, x, y, fill = "tozeroy", color = "red", shape = "spline")
}
\description{
面积图
}
\examples{
mtcars |>
  count(cyl) |>
  plot_area(x = "cyl", y = "n")
  
mtcars |> group_by(cyl) |>
  summarise(displ = mean(disp)) |>
  plot_area(x = "cyl", y = "displ")
  
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{lib-plotly functions}
