% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{plot_bar}
\alias{plot_bar}
\title{柱图}
\usage{
plot_bar(d, x, y)
}
\description{
柱图
}
\examples{
mtcars |>
  count(cyl) |>
  plot_bar(x = "cyl", y = "n")

mpg |> count(class) |>
  mutate(class = forcats::fct_reorder(class, n, .desc = TRUE)) |>
  plot_bar(x = "class", y = "n")
}
\seealso{
Other task-lib functions: 
\code{\link{gali_DT}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{task-lib functions}
