% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{plot_barpolar}
\alias{plot_barpolar}
\title{极坐标柱}
\usage{
plot_barpolar(
  d,
  theta,
  r,
  color = NULL,
  bargap = 0,
  hole = 0.05,
  direction = "clockwise"
)
}
\description{
极坐标柱
}
\examples{
mtcars |>
  count(cyl) |>
  mutate(cyl = sprintf("CYL: \%d", cyl)) |>
  plot_barpolar(theta = "cyl", r = "n")
  
}
\seealso{
Other task-lib functions: 
\code{\link{gali_DT}()},
\code{\link{plot_area}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{task-lib functions}
