% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_hist}
\alias{plot_hist}
\title{直方图}
\usage{
plot_hist(d, e_x, c_color = "brown", s_xtitle = NULL, s_ytitle = NULL)
}
\description{
直方图
}
\examples{
mtcars |> gali_plot_hist(e_x = "mpg")
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot0}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
