% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-plotly.R
\name{plot_line}
\alias{plot_line}
\title{线图}
\usage{
plot_line(
  d,
  x,
  y,
  fillcolor = "red",
  linecolor = "rgb(205, 12, 24)",
  shape = "spline"
)
}
\description{
线图
}
\examples{
mtcars |>
  count(cyl) |> 
  plot_line(x = "cyl", y = "n")

}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{lib-plotly functions}
