% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_line}
\alias{plot_line}
\title{线图}
\usage{
plot_line(
  d,
  e_x,
  e_y,
  e_fillcolor = "red",
  c_linecolor = "rgb(205, 12, 24)",
  e_shape = "spline"
)
}
\description{
线图
}
\examples{
mtcars |>
  count(cyl) |> 
  gali_plot_line(e_x = "cyl", e_y = "n")

}
\seealso{
Other lib-plotly functions: 
\code{\link{plot0}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_layout}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
