% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{plot_marker}
\alias{plot_marker}
\title{散点图}
\usage{
plot_marker(d, x, y, alpha = 0.2, name = NULL)
}
\description{
散点图
}
\examples{
mtcars |>
 plot_marker(x = "mpg", y = "disp", alpha = 0.6)
}
\seealso{
Other task-lib functions: 
\code{\link{gali_DT}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{task-lib functions}
