% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_pie}
\alias{plot_pie}
\title{饼图}
\usage{
plot_pie(d, e_value, s_label = NULL, f1_hole = 0.4)
}
\description{
饼图
}
\examples{
mtcars |> gali_plot_pie(e_value = "cyl")
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_init}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
