% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-lib-plotly.R
\name{plot_pie}
\alias{plot_pie}
\title{饼图}
\usage{
plot_pie(d, value, label = NULL, hole = 0.4)
}
\description{
饼图
}
\examples{
mtcars |> plot_pie(value = "cyl")
}
\seealso{
Other task-lib functions: 
\code{\link{gali_DT}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()}
}
\concept{task-lib functions}
