% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_pie2}
\alias{plot_pie2}
\title{饼图}
\usage{
plot_pie2(d, e_value, s_label = NULL, f1s_pull = 0, f1_hole = 0.25)
}
\description{
饼图
}
\examples{
mtcars |> count(cyl) |>
 gali_plot_pie2(e_value = "n", pull = c(0.1, 0, 0))
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_init}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie}()},
\code{\link{plot_trace}()}
}
\concept{lib-plotly functions}
