% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs-plotly.R
\name{plot_trace}
\alias{plot_trace}
\title{绘制坐标轴}
\usage{
plot_trace(plot, type, ...)
}
\description{
绘制坐标轴
}
\seealso{
Other lib-plotly functions: 
\code{\link{plot0}()},
\code{\link{plot_area}()},
\code{\link{plot_barpolar}()},
\code{\link{plot_bar}()},
\code{\link{plot_hist}()},
\code{\link{plot_layout}()},
\code{\link{plot_line}()},
\code{\link{plot_marker}()},
\code{\link{plot_pie2}()},
\code{\link{plot_pie}()}
}
\concept{lib-plotly functions}
