% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{queue_batch_todo}
\alias{queue_batch_todo}
\title{待执行的队列任务}
\usage{
queue_batch_todo(
  taskType = NULL,
  dsName = "__TASK_QUEUE__",
  cacheTopic = "CACHE"
)
}
\description{
待执行的队列任务
}
\seealso{
Other task function: 
\code{\link{queue_batch_run}()},
\code{\link{queue_dataset_init}()},
\code{\link{queue_task_item}()}
}
\concept{task function}
