% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_clear}
\alias{risk_data_clear}
\title{清理未处理的疑点数据}
\usage{
risk_data_clear(
  modelId,
  riskDataName = "__RISK_DATA__",
  cacheTopic = "CACHE",
  taskTopic = "TASK_DEFINE"
)
}
\description{
重新生成模型时，一般需要清理未处理的疑点数据
}
\seealso{
Other risk function: 
\code{\link{risk_data_build}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_read}()},
\code{\link{risk_model_create}()}
}
\concept{risk function}
