% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_init}
\alias{risk_data_init}
\title{初始化风险筛查数据集}
\usage{
risk_data_init(dsName = "__RISK_DATA__", cacheTopic = "CACHE")
}
\description{
通过风险模型筛查疑点数据
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_read}()},
\code{\link{risk_data_write}()},
\code{\link{risk_model_create}()},
\code{\link{risk_model_search}()}
}
\concept{risk function}
