% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_read}
\alias{risk_data_read}
\title{读取疑点数据}
\usage{
risk_data_read(
  todoFlag = TRUE,
  riskDataName = "__RISK_DATA__",
  cacheTopic = "CACHE"
)
}
\description{
读取疑点数据
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_write}()},
\code{\link{risk_model_create}()},
\code{\link{risk_model_search}()}
}
\concept{risk function}
