% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_data_write}
\alias{risk_data_write}
\title{写入疑点数据}
\usage{
risk_data_write(
  d,
  dsName,
  modelId,
  riskName = "__RISK_DATA__",
  cacheTopic = "CACHE",
  taskTopic = "TASK_DEFINE"
)
}
\description{
写入疑点数据
}
