% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_create}
\alias{risk_model_create}
\title{定义风险模型}
\usage{
risk_model_create(
  dsName,
  modelName,
  tagName = "main",
  riskTip = "-",
  riskLevel = "L",
  modelDesc = "-",
  author = "-",
  force = FALSE,
  riskDataName = "__RISK_DATA__",
  cacheTopic = "CACHE",
  taskTopic = "TASK_DEFINE",
  scriptsTopic = "TASK_SCRIPTS"
)
}
\description{
风险模型属于任务的一种。

风险模型定义后，主要支持对数据集做过滤查询。
}
\seealso{
Other risk function: 
\code{\link{risk_data_clear}()},
\code{\link{risk_data_init}()},
\code{\link{risk_data_read}()},
\code{\link{risk_data_set_done}()},
\code{\link{risk_data_set_todo}()},
\code{\link{risk_data_write}()},
\code{\link{risk_model_search}()}
}
\concept{risk function}
