% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_create}
\alias{risk_model_create}
\title{初始化风险模型}
\usage{
risk_model_create(
  modelName,
  dataset,
  filter = list(list(column = "col_name1", op = ">", value = 0:1, riskTip = "-", level
    = 1)),
  modelGroup = modelName,
  modelDesc = "-",
  author = "-",
  online = FALSE,
  overwrite = FALSE,
  topic = "RISKMODEL"
)
}
\arguments{
\item{modelName}{模型名称}

\item{dataset}{筛查目标数据集（该数据集必须存在yml配置文件，且设置了关键列）}

\item{filter}{筛查条件}

\item{modelGroup}{模型组}

\item{modelDesc}{模型描述}

\item{author}{模型作者}

\item{online}{是否启用}

\item{overwrite}{覆盖旧有模型}

\item{topic}{风险模型的主题域}
}
\description{
初始化风险模型
}
\seealso{
Other risk function: 
\code{\link{risk_data_read}()},
\code{\link{risk_model_all}()},
\code{\link{risk_model_read}()},
\code{\link{risk_model_run}()}
}
\concept{risk function}
