% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk_model_run}
\alias{risk_model_run}
\title{执行模型}
\usage{
risk_model_run(
  dataset = "疑点数据",
  targetTopic = "CACHE",
  topic = "RISKMODEL"
)
}
\description{
输出风险疑点数据到CACHE目录中
}
\seealso{
Other risk function: 
\code{\link{risk_data_read}()},
\code{\link{risk_model_create}()},
\code{\link{risk_model_read}()}
}
\concept{risk function}
