% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{script_item_exchange}
\alias{script_item_exchange}
\title{调整任务脚本顺序}
\usage{
script_item_exchange(taskName, rowNumA, rowNumB, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskName}{任务标识}

\item{taskTopic}{任务主题存储位置}

\item{rowNumOld}{脚本旧序号}

\item{rowNumNew}{脚本新序号}
}
\description{
调整任务脚本顺序
}
\seealso{
Other task-script function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_remove}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()}
}
\concept{task-script function}
