% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks-folder.R
\name{set_topic}
\alias{set_topic}
\title{设置主题目录}
\usage{
set_topic(topic, path)
}
\arguments{
\item{topic}{任务主题}

\item{path}{文件夹位置}
}
\description{
设置主题目录
}
\seealso{
Other TaskFolder functions: 
\code{\link{ds_import}()},
\code{\link{get_path}()},
\code{\link{get_task_scripts}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{run_task_scripts}()}
}
\concept{TaskFolder functions}
