% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{set_topic}
\alias{set_topic}
\title{设置主题目录}
\usage{
set_topic(topic, path)
}
\arguments{
\item{topic}{任务主题}

\item{path}{文件夹位置}
}
\description{
设置主题目录
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{load_config}()},
\code{\link{task_dir}()},
\code{\link{task_files}()},
\code{\link{task_run}()}
}
\concept{task functions}
