% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_create}
\alias{task_create}
\title{创建任务执行框架}
\usage{
task_create(
  taskId,
  runLevel = 500L,
  online = TRUE,
  taskType = "__UNKNOWN__",
  desc = "-",
  taskTopic = "TASK_DEFINE"
)
}
\arguments{
\item{taskId}{任务唯一标识，每个\code{taskId}会保存为一个独立文件}

\item{runLevel}{同一批次任务中，运行时的优先级}

\item{online}{任务是否启用}

\item{taskType}{任务类型}

\item{desc}{任务描述}

\item{taskTopic}{保存任务定义的存储主题文件夹}
}
\description{
创建任务执行框架
}
\seealso{
Other task-define function: 
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_script_dir_create}()},
\code{\link{task_script_file_create}()},
\code{\link{task_search}()}
}
\concept{task-define function}
