% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_create}
\alias{task_create}
\title{创建任务执行框架}
\usage{
task_create(
  taskId,
  online = FALSE,
  items = NULL,
  taskType = "__UNKNOWN__",
  desc = "-",
  taskTopic = "TASK_DEFINE",
  scriptsTopic = "TASK_SCRIPTS",
  queueDataset = "__TASK_QUEUE__",
  importDataset = "__IMPORT_FILES__",
  cacheTopic = "CACHE",
  importTopic = "IMPORT",
  extention = list()
)
}
\arguments{
\item{taskId}{任务唯一标识，每个\code{taskId}会保存为一个独立文件}

\item{online}{如果任务下线，在推荐可执行任务时将忽略}

\item{taskType}{任务类型}

\item{desc}{任务描述}

\item{taskTopic}{保存任务定义的存储主题文件夹}

\item{runLevel}{同一批次任务中，运行时的优先级}
}
\description{
创建任务执行框架
}
\seealso{
Other task-define function: 
\code{\link{task_dir_add}()},
\code{\link{task_empty_add}()},
\code{\link{task_expr_add}()},
\code{\link{task_file_add}()},
\code{\link{task_func_add}()},
\code{\link{task_gali_add}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()}
}
\concept{task-define function}
