% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-define.R
\name{task_create}
\alias{task_create}
\title{创建任务执行框架}
\usage{
task_create(
  taskId,
  items = tibble(),
  runLevel = 500,
  online = TRUE,
  taskType = "__UNKNOWN__",
  desc = "-",
  taskTopic = "TASK_DEFINE"
)
}
\arguments{
\item{taskId}{任务唯一标识，每个\code{taskId}会保存为一个独立文件}

\item{items}{子任务清单，包含子任务的执行脚本、参数设定等}

\item{runLevel}{同一批次任务中，运行时的优先级}

\item{online}{任务是否启用}

\item{taskType}{任务类型}

\item{desc}{任务描述}

\item{taskTopic}{保存任务定义的存储主题文件夹}

\item{family}{task-define function}
}
\description{
创建任务执行框架
}
