% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_id}
\alias{task_id}
\title{查询任务ID}
\usage{
task_id(taskMatch = ".*", taskTopic = "TASK_DEFINE")
}
\description{
查询任务ID
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_item_gali_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_dp}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
