% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_id}
\alias{task_id}
\title{查询任务ID}
\usage{
task_id(taskMatch = ".*", taskTopic = "TASK_DEFINE")
}
\description{
查询任务ID
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_empty_add}()},
\code{\link{task_expr_add}()},
\code{\link{task_file_add}()},
\code{\link{task_func_add}()},
\code{\link{task_item_add}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()}
}
\concept{task-define function}
