% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_item_add}
\alias{task_item_add}
\title{增加子任务}
\usage{
task_item_add(
  taskId,
  taskScript = "ls()",
  params = list(NULL),
  inputAsign = list(NULL),
  outputAsign = list(NULL),
  touchFiles = TRUE,
  taskTopic = "TASK_DEFINE",
  scriptType = "empty"
)
}
\arguments{
\item{taskId}{任务标识}

\item{taskScript}{子任务的执行路径}

\item{params}{执行脚本的参数映射}

\item{inputAsign}{将执行环境内变量赋值给入参}

\item{outputAsign}{将函数输出复制该执行环境内的变量}

\item{taskTopic}{任务主题存储位置}

\item{scriptType}{可以是string,expr,function,gali,file,dir,empty等}
}
\description{
增加子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_gali_add}()},
\code{\link{task_id}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
