% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_item_add}
\alias{task_item_add}
\title{增加子任务}
\usage{
task_item_add(
  taskId,
  script = NA,
  params = NA,
  globalVars = NA,
  inputAsign = NA,
  outputAsign = NA,
  touchFiles = TRUE,
  taskTopic = "TASK_DEFINE",
  type = "global"
)
}
\arguments{
\item{taskId}{任务标识}

\item{script}{执行脚本或函数名、文件名、目录名}

\item{params}{执行函数参数映射}

\item{globalVars}{设置全局变量}

\item{inputAsign}{针对function和gali类型，使用执行环境内变量映射入参}

\item{outputAsign}{保存子任务输出}

\item{touchFiles}{类型为file,dir时自动创建脚本文件}

\item{taskTopic}{任务主题存储位置}

\item{type}{可以是string,expr,function,gali,file,dir,global等}
}
\description{
增加子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_dir_update}()},
\code{\link{task_expr_add}()},
\code{\link{task_expr_update}()},
\code{\link{task_file_add}()},
\code{\link{task_file_update}()},
\code{\link{task_func_add}()},
\code{\link{task_func_update}()},
\code{\link{task_global_add}()},
\code{\link{task_global_update}()},
\code{\link{task_item_exchange}()},
\code{\link{task_item_remove}()},
\code{\link{task_item_update}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_string_add}()},
\code{\link{task_string_update}()},
\code{\link{task_update}()}
}
\concept{task-define function}
