% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_item_add}
\alias{task_item_add}
\title{增加子任务}
\usage{
task_item_add(
  taskId,
  taskScript,
  params = list(NULL),
  scriptType = "string",
  taskTopic = "TASK_DEFINE"
)
}
\arguments{
\item{taskId}{任务标识}

\item{taskScript}{子任务的执行路径}

\item{params}{子任务的参数设置}

\item{scriptType}{可以是string,file,dir, 或其他系统内置函数}

\item{taskTopic}{保存任务定义的存储主题文件夹}
}
\description{
增加子任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_gali_add}()},
\code{\link{task_id}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run_gali}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
