% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_online}
\alias{task_online}
\title{任务上线}
\usage{
task_online(...)
}
\description{
任务上线
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_exists}()},
\code{\link{task_id}()},
\code{\link{task_name}()},
\code{\link{task_offline}()},
\code{\link{task_params_assign}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
