% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-queue.R
\name{task_queue_item}
\alias{task_queue_item}
\title{构造队列中的一条数据}
\usage{
task_queue_item(
  taskName,
  yamlParams = "[]\\n",
  id = gen_batchNum(),
  taskTopic = "TASK_DEFINE",
  cacheTopic = "CACHE"
)
}
\description{
构造队列中的一条数据
}
\seealso{
Other queue function: 
\code{\link{task_queue_init}()},
\code{\link{task_queue_search}()}
}
\concept{queue function}
