% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_read}
\alias{task_read}
\title{读取任务}
\usage{
task_read(taskId, snap = FALSE, taskTopic = "TASK_DEFINE", snapTopic = "SNAP")
}
\arguments{
\item{taskId}{任务标识}

\item{taskTopic}{保存任务定义的存储主题文件夹}
}
\description{
读取任务
}
\seealso{
Other task-define function: 
\code{\link{script_dir_add}()},
\code{\link{script_dir_update}()},
\code{\link{script_expr_add}()},
\code{\link{script_expr_update}()},
\code{\link{script_file_add}()},
\code{\link{script_file_update}()},
\code{\link{script_func_add}()},
\code{\link{script_func_update}()},
\code{\link{script_item_add}()},
\code{\link{script_item_exchange}()},
\code{\link{script_item_remove}()},
\code{\link{script_item_update}()},
\code{\link{script_string_add}()},
\code{\link{script_string_update}()},
\code{\link{script_var_add}()},
\code{\link{script_var_update}()},
\code{\link{task_create}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
