% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_read}
\alias{task_read}
\title{读取任务}
\usage{
task_read(taskId, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskId}{任务标识}

\item{taskTopic}{保存任务定义的存储主题文件夹}

\item{family}{task-define function}
}
\description{
读取任务
}
