% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_read}
\alias{task_read}
\title{读取任务}
\usage{
task_read(taskId, taskTopic = "TASK_DEFINE")
}
\arguments{
\item{taskId}{任务标识}

\item{taskTopic}{保存任务定义的存储主题文件夹}
}
\description{
读取任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_expr}()},
\code{\link{task_run_file}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_script_dir_create}()},
\code{\link{task_script_file_create}()},
\code{\link{task_search}()}
}
\concept{task-define function}
