% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_read}
\alias{task_read}
\title{读取任务}
\usage{
task_read(
  taskName,
  snap = FALSE,
  taskTopic = "TASK_DEFINE",
  snapTopic = "SNAP"
)
}
\arguments{
\item{taskName}{任务标识}

\item{taskTopic}{保存任务定义的存储主题文件夹}
}
\description{
读取任务
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_exists}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_run}()},
\code{\link{task_search}()},
\code{\link{task_update}()}
}
\concept{task-define function}
