% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run}
\alias{task_run}
\title{执行目标路径下的任务脚本}
\usage{
task_run(
  taskTopic = "TASK/BUILD",
  taskFolder = "",
  importTopic = "IMPORT",
  batchNum = gen_batchNum(),
  glob = "*.R"
)
}
\arguments{
\item{taskTopic}{脚本文件夹主题}

\item{taskFolder}{执行脚本文件的目录}

\item{importTopic}{导入主题域}

\item{glob}{默认加载所有R文件}
}
\description{
应当按照脚本顺序执行
}
\details{
glob参数可用于运行运行特定的R文件，
例如 \code{task_run(taskFolder = "abc", glob = "**/1.R")}。
}
\seealso{
Other task functions: 
\code{\link{get_importing_folder}()},
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{import_folders}()},
\code{\link{import_redo}()},
\code{\link{import_todo}()},
\code{\link{set_topic}()},
\code{\link{task_dir}()},
\code{\link{task_files}()}
}
\concept{task functions}
