% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run}
\alias{task_run}
\title{执行目标路径下的任务脚本}
\usage{
task_run(taskScript = "TASK/BUILD", batch = F)
}
\description{
应当按照脚本顺序执行
}
\seealso{
Other TaskFolder functions: 
\code{\link{get_path}()},
\code{\link{get_topics}()},
\code{\link{get_topic}()},
\code{\link{set_topic}()},
\code{\link{task_read}()}
}
\concept{TaskFolder functions}
