% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run}
\alias{task_run}
\title{运行任务}
\usage{
task_run(
  taskId,
  withQueue = FALSE,
  taskTopic = "TASK_DEFINE",
  runMode = "in-process",
  ...
)
}
\arguments{
\item{taskId}{任务标识}

\item{withQueue}{是否在运行队列中显示状态}

\item{taskTopic}{保存任务定义的存储主题文件夹}

\item{runMode}{运行模式（默认为进程内执行，改为r或r_bg为子进程执行）}
}
\description{
运行任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_gali_add}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_search}()}
}
\concept{task-define function}
