% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run_expr}
\alias{task_run_expr}
\title{快速运行表达式}
\usage{
task_run_expr(taskExpr, params = list(NULL), runMode = "in-process", ...)
}
\arguments{
\item{taskExpr}{脚本表达式}
}
\description{
快速运行表达式
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_gali_add}()},
\code{\link{task_id}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_file}()},
\code{\link{task_run_gali}()},
\code{\link{task_run_string}()},
\code{\link{task_run}()},
\code{\link{task_search}()}
}
\concept{task-define function}
