% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_run_string}
\alias{task_run_string}
\title{运行任务}
\usage{
task_run_string(taskString, params = list(NULL), runMode = "in-process", ...)
}
\arguments{
\item{taskFile}{任务文件}
}
\description{
运行任务
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_item_add}()},
\code{\link{task_read}()},
\code{\link{task_run0}()},
\code{\link{task_run_dir}()},
\code{\link{task_run_file}()},
\code{\link{task_run}()},
\code{\link{task_script_dir_create}()},
\code{\link{task_script_file_create}()},
\code{\link{task_search}()}
}
\concept{task-define function}
