% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_search}
\alias{task_search}
\title{列举所有任务定义}
\usage{
task_search(taskMatch = ".*", typeMatch = ".*", taskTopic = "TASK_DEFINE")
}
\arguments{
\item{topic}{主题域}
}
\description{
列举所有任务定义
}
\seealso{
Other task-define function: 
\code{\link{task_cancel_snap}()},
\code{\link{task_create}()},
\code{\link{task_offline}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_update}()}
}
\concept{task-define function}
