% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_search}
\alias{task_search}
\title{列举所有任务定义}
\usage{
task_search(taskMatch = ".*", typeMatch = ".*", taskTopic = "TASK_DEFINE")
}
\arguments{
\item{topic}{主题域}
}
\description{
列举所有任务定义
}
\seealso{
Other task-define function: 
\code{\link{task_create}()},
\code{\link{task_dir_add}()},
\code{\link{task_dir_update}()},
\code{\link{task_expr_add}()},
\code{\link{task_expr_update}()},
\code{\link{task_file_add}()},
\code{\link{task_file_update}()},
\code{\link{task_func_add}()},
\code{\link{task_func_update}()},
\code{\link{task_global_add}()},
\code{\link{task_global_update}()},
\code{\link{task_item_add}()},
\code{\link{task_item_exchange}()},
\code{\link{task_item_remove}()},
\code{\link{task_item_update}()},
\code{\link{task_online}()},
\code{\link{task_read}()},
\code{\link{task_run}()},
\code{\link{task_string_add}()},
\code{\link{task_string_update}()},
\code{\link{task_update}()}
}
\concept{task-define function}
